bitcoin::address

Struct Address

pub struct Address<V = NetworkChecked>
where
    V: NetworkValidation,
{
    pub payload: Payload,
    pub network: Network,
    /* private fields */
}

A Bitcoin address.

Parsing addresses

When parsing string as an address, one has to pay attention to the network, on which the parsed address is supposed to be valid. For the purpose of this validation, Address has is_valid_for_network method. In order to provide more safety, enforced by compiler, Address also contains a special marker type, which indicates whether network of the parsed address has been checked. This marker type will prevent from calling certain functions unless the network verification has been successfully completed.

The result of parsing an address is Address<NetworkUnchecked> suggesting that network of the parsed address has not yet been verified. To perform this verification, method require_network can be called, providing network on which the address is supposed to be valid. If the verification succeeds, Address<NetworkChecked> is returned.

The types Address and Address<NetworkChecked> are synonymous, i. e. they can be used interchangeably.

use std::str::FromStr;
use bitcoin::{Address, Network};
use bitcoin::address::{NetworkUnchecked, NetworkChecked};

// variant 1
let address: Address<NetworkUnchecked> = "32iVBEu4dxkUQk9dJbZUiBiQdmypcEyJRf".parse().unwrap();
let address: Address<NetworkChecked> = address.require_network(Network::Bitcoin).unwrap();

// variant 2
let address: Address = Address::from_str("32iVBEu4dxkUQk9dJbZUiBiQdmypcEyJRf").unwrap()
               .require_network(Network::Bitcoin).unwrap();

// variant 3
let address: Address<NetworkChecked> = "32iVBEu4dxkUQk9dJbZUiBiQdmypcEyJRf".parse::<Address<_>>()
               .unwrap().require_network(Network::Bitcoin).unwrap();

Formatting addresses

To format address into its textual representation, both Debug (for usage in programmer-facing, debugging context) and Display (for user-facing output) can be used, with the following caveats:

Display is implemented only for Address<NetworkChecked>:

let address: Address<NetworkChecked> = Address::from_str("132F25rTsvBdp9JzLLBHP5mvGY66i1xdiM")
               .unwrap().assume_checked();
assert_eq!(address.to_string(), "132F25rTsvBdp9JzLLBHP5mvGY66i1xdiM");

let address: Address<NetworkUnchecked> = Address::from_str("132F25rTsvBdp9JzLLBHP5mvGY66i1xdiM")
               .unwrap();
let s = address.to_string(); // does not compile

Debug on Address<NetworkUnchecked> does not produce clean address but address wrapped by an indicator that its network has not been checked. This is to encourage programmer to properly check the network and use Display in user-facing context.

let address: Address<NetworkUnchecked> = Address::from_str("132F25rTsvBdp9JzLLBHP5mvGY66i1xdiM")
               .unwrap();
assert_eq!(format!("{:?}", address), "Address<NetworkUnchecked>(132F25rTsvBdp9JzLLBHP5mvGY66i1xdiM)");

let address: Address<NetworkChecked> = Address::from_str("132F25rTsvBdp9JzLLBHP5mvGY66i1xdiM")
               .unwrap().assume_checked();
assert_eq!(format!("{:?}", address), "132F25rTsvBdp9JzLLBHP5mvGY66i1xdiM");

Relevant BIPs

BIP13 - Address Format for pay-to-script-hash

BIP16 - Pay to Script Hash

BIP141 - Segregated Witness (Consensus layer)

BIP142 - Address Format for Segregated Witness

BIP341 - Taproot: SegWit version 1 spending rules

BIP350 - Bech32m format for v1+ witness addresses

Fields
payload: Payload

The type of the address.

network: Network

The network on which this address is usable.

Implementations

impl<V: NetworkValidation> Address<V>

Methods on Address that can be called on both Address<NetworkChecked> and Address<NetworkUnchecked>.

pub fn new(network: Network, payload: Payload) -> Address<V>

Create new address from given components, infering the network validation marker type of the address.

impl Address

Methods and functions that can be called only on Address<NetworkChecked>.

pub fn p2pkh(pk: &PublicKey, network: Network) -> Address

Creates a pay to (compressed) public key hash address from a public key.

This is the preferred non-witness type address.

pub fn p2sh(script: &Script, network: Network) -> Result<Address, Error>

Creates a pay to script hash P2SH address from a script.

This address type was introduced with BIP16 and is the popular type to implement multi-sig these days.

pub fn p2wpkh(pk: &PublicKey, network: Network) -> Result<Address, Error>

Creates a witness pay to public key address from a public key.

This is the native segwit address type for an output redeemable with a single signature.

Errors

Will only return an error if an uncompressed public key is provided.

pub fn p2shwpkh(pk: &PublicKey, network: Network) -> Result<Address, Error>

Creates a pay to script address that embeds a witness pay to public key.

This is a segwit address type that looks familiar (as p2sh) to legacy clients.

Errors

Will only return an Error if an uncompressed public key is provided.

pub fn p2wsh(script: &Script, network: Network) -> Address

Creates a witness pay to script hash address.

pub fn p2shwsh(script: &Script, network: Network) -> Address

Creates a pay to script address that embeds a witness pay to script hash address.

This is a segwit address type that looks familiar (as p2sh) to legacy clients.

pub fn p2tr<C: Verification>( secp: &Secp256k1<C>, internal_key: UntweakedPublicKey, merkle_root: Option<TapNodeHash>, network: Network, ) -> Address

Creates a pay to taproot address from an untweaked key.

pub fn p2tr_tweaked(output_key: TweakedPublicKey, network: Network) -> Address

Creates a pay to taproot address from a pre-tweaked output key.

This method is not recommended for use, Address::p2tr() should be used where possible.

pub fn address_type(&self) -> Option<AddressType>

Gets the address type of the address.

Returns

None if unknown, non-standard or related to the future witness version.

pub fn is_spend_standard(&self) -> bool

Checks whether or not the address is following Bitcoin standardness rules when spending from this address. NOT to be called by senders.

Spending Standardness

For forward compatibility, the senders must send to any Address. Receivers can use this method to check whether or not they can spend from this address.

SegWit addresses with unassigned witness versions or non-standard program sizes are considered non-standard.

pub fn is_standard(&self) -> bool

👎Deprecated since 0.30.0: Use Address::is_spend_standard instead

Checks whether or not the address is following Bitcoin standardness rules.

SegWit addresses with unassigned witness versions or non-standard program sizes are considered non-standard.

pub fn from_script(script: &Script, network: Network) -> Result<Address, Error>

Constructs an Address from an output script (scriptPubkey).

pub fn script_pubkey(&self) -> ScriptBuf

Generates a script pubkey spending to this address.

pub fn to_qr_uri(&self) -> String

Creates a URI string bitcoin:address optimized to be encoded in QR codes.

If the address is bech32, both the schema and the address become uppercase. If the address is base58, the schema is lowercase and the address is left mixed case.

Quoting BIP 173 “inside QR codes uppercase SHOULD be used, as those permit the use of alphanumeric mode, which is 45% more compact than the normal byte mode.”

Note however that despite BIP21 explicitly stating that the bitcoin: prefix should be parsed as case-insensitive many wallets got this wrong and don’t parse correctly. See compatibility table.

If you want to avoid allocation you can use alternate display instead:

write!(writer, "{:#}", address)?;

pub fn is_related_to_pubkey(&self, pubkey: &PublicKey) -> bool

Returns true if the given pubkey is directly related to the address payload.

This is determined by directly comparing the address payload with either the hash of the given public key or the segwit redeem hash generated from the given key. For taproot addresses, the supplied key is assumed to be tweaked

pub fn is_related_to_xonly_pubkey(&self, xonly_pubkey: &XOnlyPublicKey) -> bool

Returns true if the supplied xonly public key can be used to derive the address.

This will only work for Taproot addresses. The Public Key is assumed to have already been tweaked.

pub fn matches_script_pubkey(&self, script_pubkey: &Script) -> bool

Returns true if the address creates a particular script This function doesn’t make any allocations.

impl Address<NetworkUnchecked>

Methods that can be called only on Address<NetworkUnchecked>.

pub fn is_valid_for_network(&self, network: Network) -> bool

Parsed addresses do not always have one network. The problem is that legacy testnet, regtest and signet addresse use the same prefix instead of multiple different ones. When parsing, such addresses are always assumed to be testnet addresses (the same is true for bech32 signet addresses). So if one wants to check if an address belongs to a certain network a simple comparison is not enough anymore. Instead this function can be used.

use bitcoin::{Address, Network};
use bitcoin::address::NetworkUnchecked;

let address: Address<NetworkUnchecked> = "2N83imGV3gPwBzKJQvWJ7cRUY2SpUyU6A5e".parse().unwrap();
assert!(address.is_valid_for_network(Network::Testnet));
assert!(address.is_valid_for_network(Network::Regtest));
assert!(address.is_valid_for_network(Network::Signet));

assert_eq!(address.is_valid_for_network(Network::Bitcoin), false);

let address: Address<NetworkUnchecked> = "32iVBEu4dxkUQk9dJbZUiBiQdmypcEyJRf".parse().unwrap();
assert!(address.is_valid_for_network(Network::Bitcoin));
assert_eq!(address.is_valid_for_network(Network::Testnet), false);

pub fn require_network(self, required: Network) -> Result<Address, Error>

Checks whether network of this address is as required.

For details about this mechanism, see section Parsing addresses on Address.

pub fn assume_checked(self) -> Address

Marks, without any additional checks, network of this address as checked.

Improper use of this method may lead to loss of funds. Reader will most likely prefer require_network as a safe variant. For details about this mechanism, see section Parsing addresses on Address.