bitcoin

Module amount

Bitcoin amounts.

This module mainly introduces the Amount and SignedAmount types. We refer to the documentation on the types for more information.

Structs

Amount

Amount

Display

A helper/builder that displays amount with specified settings.

SignedAmount

SignedAmount

Enums

Denomination

A set of denominations in which amounts can be expressed.

ParseAmountError

An error during amount parsing.

Traits

CheckedSum

Calculate the sum over the iterator using checked arithmetic.