bitcoin::amount::serde
Trait SerdeAmount

pub trait SerdeAmount:
    Copy
    + Sized
    + Sealed {
    // Required methods
    fn ser_sat<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>;
    fn des_sat<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>;
    fn ser_btc<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>;
    fn des_btc<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>;
}

Available on crate feature serde only.

This trait is used only to avoid code duplication and naming collisions of the different serde serialization crates.

Required Methods

fn ser_sat<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>

fn des_sat<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>

fn ser_btc<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>

fn des_btc<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>

Object Safety

This trait is not object safe.