bitcoin::amount

Trait CheckedSum

pub trait CheckedSum<R>: SumSeal<R> {
    // Required method
    fn checked_sum(self) -> Option<R>;
}

Calculate the sum over the iterator using checked arithmetic.

Required Methods

fn checked_sum(self) -> Option<R>

Calculate the sum over the iterator using checked arithmetic. If an over or underflow would happen it returns None.