bitcoin
Module base58

Base58 encoder and decoder.

This module provides functions for encoding and decoding base58 slices and strings respectively.

Enums

Error

An error that might occur during base58 decoding.

Functions

decode

Decodes a base58-encoded string into a byte vector.

decode_check

Decodes a base58check-encoded string into a byte vector verifying the checksum.

encode

Encodes data as a base58 string (see also base58::encode_check()).

encode_check

Encodes data as a base58 string including the checksum.

encode_check_to_fmt

Encodes a slice as base58, including the checksum, into a formatter.