bitcoin::bip152

Struct ShortId

pub struct ShortId(/* private fields */);

Short transaction IDs are used to represent a transaction without sending a full 256-bit hash.

Implementations

impl ShortId

pub fn as_ptr(&self) -> *const u8

Converts the object to a raw pointer.

pub fn as_mut_ptr(&mut self) -> *mut u8

Converts the object to a mutable raw pointer.

pub fn len(&self) -> usize

Returns the length of the object as an array.

pub fn is_empty(&self) -> bool

Returns whether the object, as an array, is empty. Always false.

impl ShortId

pub fn as_bytes(&self) -> &[u8; 6]

Returns a reference the underlying bytes.

pub fn to_bytes(self) -> [u8; 6]

Returns the underlying bytes.

impl ShortId

pub fn calculate_siphash_keys(header: &Header, nonce: u64) -> (u64, u64)

Calculate the SipHash24 keys used to calculate short IDs.

pub fn with_siphash_keys<T: AsRef<[u8]>>( txid: &T, siphash_keys: (u64, u64), ) -> ShortId

Calculate the short ID with the given (w)txid and using the provided SipHash keys.