bitcoin::bip32
Type Alias ExtendendPrivKey

pub type ExtendendPrivKey = Xpriv;

👎Deprecated since 0.31.0: use xpriv instead

The old name for xpriv, extended public key.

Aliased Type

struct ExtendendPrivKey {
    pub network: Network,
    pub depth: u8,
    pub parent_fingerprint: Fingerprint,
    pub child_number: ChildNumber,
    pub private_key: SecretKey,
    pub chain_code: ChainCode,
}

Fields
network: Network

The network this key is to be used on

depth: u8

How many derivations this key is from the master (which is 0)

parent_fingerprint: Fingerprint

Fingerprint of the parent key (0 for master)

child_number: ChildNumber

Child number of the key used to derive from parent (0 for master)

private_key: SecretKey

Private key

chain_code: ChainCode

Chain code