bitcoin::blockdata
Module block

Bitcoin blocks.

A block is a bundle of transactions with a proof-of-work attached, which commits to an earlier block to form the blockchain. This module describes structures and functions needed to describe these blocks and the blockchain.

Re-exports

pub use crate::hash_types::BlockHash;

Structs

Block

Bitcoin block.

Header

Bitcoin block header.

Version

Bitcoin block version number.

Enums

Bip34Error

An error when looking up a BIP34 block height.

ValidationError

A block validation error.