Struct bitcoin::blockdata::block::Block

pub struct Block {
    pub header: BlockHeader,
    pub txdata: Vec<Transaction>,
}

Bitcoin block.

A collection of transactions with an attached proof of work.

See Bitcoin Wiki: Block for more information.

Bitcoin Core References

CBlock definition

Fields
header: BlockHeader

The block header

txdata: Vec<Transaction>

List of transactions contained in the block

Implementations

impl Block

pub fn block_hash(&self) -> BlockHash

Returns the block hash.

pub fn check_merkle_root(&self) -> bool

check if merkle root of header matches merkle root of the transaction list

pub fn check_witness_commitment(&self) -> bool

Checks if witness commitment in coinbase matches the transaction list.

pub fn compute_merkle_root(&self) -> Option<TxMerkleNode>

Computes the transaction merkle root.

pub fn merkle_root(&self) -> Option<TxMerkleNode>

👎Deprecated since 0.28.0:
Please use block::compute_merkle_root instead.

Calculate the transaction merkle root.

pub fn compute_witness_commitment(
witness_root: &WitnessMerkleNode,
witness_reserved_value: &[u8]
) -> WitnessCommitment

Computes the witness commitment for the block’s transaction list.

pub fn witness_root(&self) -> Option<WitnessMerkleNode>

Computes the merkle root of transactions hashed for witness.

pub fn get_size(&self) -> usize

👎Deprecated since 0.28.0:
Please use block::size instead.

Returns the size of the block.

pub fn size(&self) -> usize

Returns the size of the block.

size == size of header + size of encoded transaction count + total size of transactions.

pub fn get_strippedsize(&self) -> usize

👎Deprecated since 0.28.0:
Please use transaction::strippedsize instead.

Returns the strippedsize of the block.

pub fn strippedsize(&self) -> usize

Returns the strippedsize of the block.

pub fn get_weight(&self) -> usize

👎Deprecated since 0.28.0:
Please use transaction::weight instead.

Returns the weight of the block.

pub fn weight(&self) -> usize

Returns the weight of the block.

pub fn coinbase(&self) -> Option<&Transaction>

Returns the coinbase transaction, if one is present.

pub fn bip34_block_height(&self) -> Result<u64, Bip34Error>

Returns the block height, as encoded in the coinbase transaction according to BIP34.