Module bitcoin::blockdata::locktime

Provides type LockTime that implements the logic around nLockTime/OP_CHECKLOCKTIMEVERIFY.

There are two types of lock time: lock-by-blockheight and lock-by-blocktime, distinguished by whether LockTime < LOCKTIME_THRESHOLD.

Structs

ConversionError

An error that occurs when converting a u32 to a lock time variant.

Height

An absolute block height, guaranteed to always contain a valid height value.

PackedLockTime

Packed lock time wraps a LockTime consensus value i.e., the raw u32 used by the network.

Time

A UNIX timestamp, seconds since epoch, guaranteed to always contain a valid time value.

Enums

Error

Catchall type for errors that relate to time locks.

LockTime

A lock time value, representing either a block height or a UNIX timestamp (seconds since epoch).

OperationError

Errors than occur when operating on lock times.

Constants

LOCK_TIME_THRESHOLD

The Threshold for deciding whether a lock time value is a height or a time (see Bitcoin Core).