bitcoin::blockdata
Module opcodes

Bitcoin script opcodes.

Bitcoin’s script uses a stack-based assembly language. This module defines all of the opcodes for that language.

Modules

all

Enables wildcard imports to bring into scope all opcodes and nothing else.

Structs

Opcode

A script Opcode.

Enums

Class

Broad categories of opcodes with similar behavior.

ClassifyContext

Classification context for the opcode.

Statics

OP_0

Push an empty array onto the stack.

OP_FALSE

Empty stack is also FALSE.

OP_NOP2

Previously called OP_NOP2.

OP_NOP3

Previously called OP_NOP3.

OP_TRUE

Number 1 is also TRUE.