Struct bitcoin::blockdata::script::PushBytes

#[repr(transparent)]
pub struct PushBytes(_);

Byte slices that can be in Bitcoin script.

The encoding of Bitcoin script restricts data pushes to be less than 2^32 bytes long. This type represents slices that are guaranteed to be within the limit so they can be put in the script safely.

Implementations

impl PushBytes

pub fn empty() -> &'static Self

Creates an empty PushBytes.

pub fn as_bytes(&self) -> &[u8]

Returns the underlying bytes.

pub fn as_mut_bytes(&mut self) -> &mut [u8]

Returns the underlying mutbale bytes.

impl PushBytes

pub fn len(&self) -> usize

Returns the number of bytes in buffer.

pub fn is_empty(&self) -> bool

Returns true if the buffer contains zero bytes.