Struct bitcoin::blockdata::script::Script

pub struct Script(_);

A Bitcoin script

Implementations

impl Script

pub fn new() -> Script

Creates a new empty script

pub fn new_p2pk(pubkey: &PublicKey) -> Script

Generates P2PK-type of scriptPubkey

pub fn new_p2pkh(pubkey_hash: &PubkeyHash) -> Script

Generates P2PKH-type of scriptPubkey

pub fn new_p2sh(script_hash: &ScriptHash) -> Script

Generates P2SH-type of scriptPubkey with a given hash of the redeem script

pub fn new_v0_wpkh(pubkey_hash: &WPubkeyHash) -> Script

Generates P2WPKH-type of scriptPubkey

pub fn new_v0_wsh(script_hash: &WScriptHash) -> Script

Generates P2WSH-type of scriptPubkey with a given hash of the redeem script

pub fn new_witness_program(ver: u5, program: &[u8]) -> Script

Generates P2WSH-type of scriptPubkey with a given hash of the redeem script

pub fn new_op_return(data: &[u8]) -> Script

Generates OP_RETURN-type of scriptPubkey for a given data

pub fn script_hash(&self) -> ScriptHash

Returns 160-bit hash of the script

pub fn wscript_hash(&self) -> WScriptHash

Returns 256-bit hash of the script for P2WSH outputs

pub fn len(&self) -> usize

The length in bytes of the script

pub fn is_empty(&self) -> bool

Whether the script is the empty script

pub fn as_bytes(&self) -> &[u8]

Returns the script data

pub fn to_bytes(&self) -> Vec<u8>

Returns a copy of the script data

pub fn into_bytes(self) -> Vec<u8>

Convert the script into a byte vector

pub fn to_p2sh(&self) -> Script

Compute the P2SH output corresponding to this redeem script

pub fn to_v0_p2wsh(&self) -> Script

Compute the P2WSH output corresponding to this witnessScript (aka the “witness redeem script”)

pub fn is_p2sh(&self) -> bool

Checks whether a script pubkey is a p2sh output

pub fn is_p2pkh(&self) -> bool

Checks whether a script pubkey is a p2pkh output

pub fn is_p2pk(&self) -> bool

Checks whether a script pubkey is a p2pk output

pub fn is_witness_program(&self) -> bool

Checks whether a script pubkey is a Segregated Witness (segwit) program.

pub fn is_v0_p2wsh(&self) -> bool

Checks whether a script pubkey is a p2wsh output

pub fn is_v0_p2wpkh(&self) -> bool

Checks whether a script pubkey is a p2wpkh output

pub fn is_op_return(&self) -> bool

Check if this is an OP_RETURN output

pub fn is_provably_unspendable(&self) -> bool

Whether a script can be proven to have no satisfying input

pub fn dust_value(&self) -> Amount

Gets the minimum value an output with this script should have in order to be broadcastable on today’s bitcoin network.

pub fn instructions(&self) -> Instructions<'_>

Iterate over the script in the form of Instructions, which are an enum covering opcodes, datapushes and errors. At most one error will be returned and then the iterator will end. To instead iterate over the script as sequence of bytes, treat it as a slice using script[..] or convert it to a vector using into_bytes().

To force minimal pushes, use Self::instructions_minimal.

pub fn instructions_minimal(&self) -> Instructions<'_>

Iterate over the script in the form of Instructions while enforcing minimal pushes.

pub fn bytes_to_asm_fmt(script: &[u8], f: &mut dyn Write) -> Result

Write the assembly decoding of the script bytes to the formatter.

pub fn fmt_asm(&self, f: &mut dyn Write) -> Result

Write the assembly decoding of the script to the formatter.

pub fn bytes_to_asm(script: &[u8]) -> String

Create an assembly decoding of the script in the given byte slice.

pub fn asm(&self) -> String

Get the assembly decoding of the script.