bitcoin::blockdata::transaction

Struct InputWeightPrediction

pub struct InputWeightPrediction { /* private fields */ }

Weight prediction of an individual input.

This helper type collects information about an input to be used in predict_weight function. It can only be created using the new function.

Implementations

impl InputWeightPrediction

pub const P2WPKH_MAX: Self

Input weight prediction corresponding to spending of P2WPKH output with the largest possible DER-encoded signature.

If the input in your transaction uses P2WPKH you can use this instead of InputWeightPrediction::new.

This is useful when you do not use signature grinding and want to ensure you are not under-paying. See ground_p2wpkh if you do use signature grinding.

pub const P2TR_KEY_DEFAULT_SIGHASH: Self

Input weight prediction corresponding to spending of taproot output using the key and default sighash.

If the input in your transaction uses Taproot key spend you can use this instead of InputWeightPrediction::new.

pub const P2TR_KEY_NON_DEFAULT_SIGHASH: Self

Input weight prediction corresponding to spending of taproot output using the key and non-default sighash.

If the input in your transaction uses Taproot key spend you can use this instead of InputWeightPrediction::new.

pub const fn ground_p2wpkh(bytes_to_grind: usize) -> Self

Input weight prediction corresponding to spending of P2WPKH output using signature grinding.

If the input in your transaction uses P2WPKH and you use signature grinding you can use this instead of InputWeightPrediction::new. See P2WPKH_MAX if you don’t use signature grinding.

Note: bytes_to_grind is usually 1 because of exponential cost of higher values.

Panics

The funcion panics in const context and debug builds if bytes_to_grind is higher than 62.

pub fn new<T>(input_script_len: usize, witness_element_lengths: T) -> Self
where T: IntoIterator, T::Item: Borrow<usize>,

Computes the prediction for a single input.

pub const fn from_slice( input_script_len: usize, witness_element_lengths: &[usize], ) -> Self

Computes the prediction for a single input in const context.

Important: only available in Rust 1.46+

This is a const version of new which only allows slices due to current Rust limitations around const fn. Because of these limitations it may be less efficient than new and thus is intended to be only used in const context.