Struct bitcoin::blockdata::transaction::Sequence

pub struct Sequence(pub u32);

Bitcoin transaction input sequence number.

The sequence field is used for:

Indicating whether absolute lock-time (specified in lock_time field of Transaction) is enabled.

Indicating and encoding BIP-68 relative lock-times.

Indicating whether a transaction opts-in to BIP-125 replace-by-fee.

Note that transactions spending an output with OP_CHECKLOCKTIMEVERIFYMUST NOT use Sequence::MAX for the corresponding input. BIP-65

Tuple Fields
0: u32
Implementations

impl Sequence

pub const MAX: Self = _

The maximum allowable sequence number.

This sequence number disables absolute lock time and replace-by-fee.

pub const ZERO: Self = _

Zero value sequence.

This sequence number enables replace-by-fee and absolute lock time.

pub const ENABLE_LOCKTIME_NO_RBF: Self = Sequence::MIN_NO_RBF

The sequence number that enables absolute lock time but disables replace-by-fee and relative lock time.

pub const ENABLE_RBF_NO_LOCKTIME: Self = _

The sequence number that enables replace-by-fee and absolute lock time but disables relative lock time.

pub const fn max_value() -> Self

👎Deprecated since 0.31.0: Use Self::MAX instead

The maximum allowable sequence number.

This is provided for consistency with Rust 1.41.1, newer code should use Sequence::MAX.

pub fn enables_absolute_lock_time(&self) -> bool

Returns true if the sequence number enables absolute lock-time (Transaction::lock_time).

pub fn is_final(&self) -> bool

Returns true if the sequence number indicates that the transaction is finalized.

Instead of this method please consider using !enables_absolute_lock_time because it is equivalent and improves readability for those not steeped in Bitcoin folklore.

Historical note

The term ‘final’ is an archaic Bitcoin term, it may have come about because the sequence number in the original Bitcoin code was intended to be incremented in order to replace a transaction, so once the sequence number got to u64::MAX it could no longer be increased, hence it was ‘final’.

Some other references to the term:

CTxIn::SEQUENCE_FINAL in the Bitcoin Core code.

BIP-112: “BIP 68 prevents a non-final transaction from being selected for inclusion in a block until the corresponding input has reached the specified age”

pub fn is_rbf(&self) -> bool

Returns true if the transaction opted-in to BIP125 replace-by-fee.

Replace by fee is signaled by the sequence being less than 0xfffffffe which is checked by this method. Note, this is the highest “non-final” value (see Sequence::is_final).

pub fn is_relative_lock_time(&self) -> bool

Returns true if the sequence has a relative lock-time.

pub fn is_height_locked(&self) -> bool

Returns true if the sequence number encodes a block based relative lock-time.

pub fn is_time_locked(&self) -> bool

Returns true if the sequence number encodes a time interval based relative lock-time.

pub fn from_height(height: u16) -> Self

Creates a relative lock-time using block height.

pub fn from_512_second_intervals(intervals: u16) -> Self

Creates a relative lock-time using time intervals where each interval is equivalent to 512 seconds.

Encoding finer granularity of time for relative lock-times is not supported in Bitcoin

pub fn from_seconds_floor(seconds: u32) -> Result<Self, Error>

Creates a relative lock-time from seconds, converting the seconds into 512 second interval with floor division.

Will return an error if the input cannot be encoded in 16 bits.

pub fn from_seconds_ceil(seconds: u32) -> Result<Self, Error>

Creates a relative lock-time from seconds, converting the seconds into 512 second interval with ceiling division.

Will return an error if the input cannot be encoded in 16 bits.

pub fn from_consensus(n: u32) -> Self

Creates a sequence from a u32 value.

pub fn to_consensus_u32(self) -> u32

Returns the inner 32bit integer value of Sequence.

pub fn to_relative_lock_time(&self) -> Option<LockTime>

Creates a relative::LockTime from this Sequence number.