Module bitcoin::blockdata::witness

Witness

This module contains the Witness struct and related methods to operate on it

Structs

Iter

Support structure to allow efficient and convenient iteration over the Witness elements

Witness

The Witness is the data used to unlock bitcoins since the segwit upgrade