Struct bitcoin::key::secp256k1::rand::rngs::adapter::ReseedingRng

pub struct ReseedingRng<R, Rsdr>(_)
where
         R: BlockRngCore + SeedableRng,
         Rsdr: RngCore;

A wrapper around any PRNG that implements BlockRngCore, that adds the ability to reseed it.

ReseedingRng reseeds the underlying PRNG in the following cases:

On a manual call to reseed().

After clone(), the clone will be reseeded on first use.

When a process is forked on UNIX, the RNGs in both the parent and child processes will be reseeded just before the next call to BlockRngCore::generate, i.e. “soon”. For ChaCha and Hc128 this is a maximum of fifteen u32 values before reseeding.

After the PRNG has generated a configurable number of random bytes.

When should reseeding after a fixed number of generated bytes be used?

Reseeding after a fixed number of generated bytes is never strictly necessary. Cryptographic PRNGs don’t have a limited number of bytes they can output, or at least not a limit reachable in any practical way. There is no such thing as ‘running out of entropy’.

Occasionally reseeding can be seen as some form of ‘security in depth’. Even if in the future a cryptographic weakness is found in the CSPRNG being used, or a flaw in the implementation, occasionally reseeding should make exploiting it much more difficult or even impossible.

Use ReseedingRng::new with a threshold of 0 to disable reseeding after a fixed number of generated bytes.

Limitations

It is recommended that a ReseedingRng (including ThreadRng) not be used from a fork handler. Use OsRng or getrandom, or defer your use of the RNG until later.

Error handling

Although unlikely, reseeding the wrapped PRNG can fail. ReseedingRng will never panic but try to handle the error intelligently through some combination of retrying and delaying reseeding until later. If handling the source error fails ReseedingRng will continue generating data from the wrapped PRNG without reseeding.

Manually calling reseed() will not have this retry or delay logic, but reports the error.

Example

use rand::prelude::*;
use rand_chacha::ChaCha20Core; // Internal part of ChaChaRng that
                             // implements BlockRngCore
use rand::rngs::OsRng;
use rand::rngs::adapter::ReseedingRng;

let prng = ChaCha20Core::from_entropy();
let mut reseeding_rng = ReseedingRng::new(prng, 0, OsRng);

println!("{}", reseeding_rng.gen::<u64>());

let mut cloned_rng = reseeding_rng.clone();
assert!(reseeding_rng.gen::<u64>() != cloned_rng.gen::<u64>());

Implementations

impl<R, Rsdr> ReseedingRng<R, Rsdr>where R: BlockRngCore + SeedableRng, Rsdr: RngCore,

pub fn new(rng: R, threshold: u64, reseeder: Rsdr) -> ReseedingRng<R, Rsdr>

Create a new ReseedingRng from an existing PRNG, combined with a RNG to use as reseeder.

threshold sets the number of generated bytes after which to reseed the PRNG. Set it to zero to never reseed based on the number of generated values.

pub fn reseed(&mut self) -> Result<(), Error>

Reseed the internal PRNG.