Struct bitcoin::key::secp256k1::XOnlyPublicKey

pub struct XOnlyPublicKey(_);

An x-only public key, used for verification of schnorr signatures and serialized according to BIP-340.

Serde support

Implements de/serialization with the serde feature enabled. We treat the byte value as a tuple of 32 u8s for non-human-readable formats. This representation is optimal for for some formats (e.g. bincode) however other formats may be less optimal (e.g. cbor).

Examples

Basic usage:

use secp256k1::{rand, Secp256k1, KeyPair, XOnlyPublicKey};

let secp = Secp256k1::new();
let key_pair = KeyPair::new(&secp, &mut rand::thread_rng());
let xonly = XOnlyPublicKey::from_keypair(&key_pair);

Implementations

impl XOnlyPublicKey

pub fn cmp_fast_unstable(&self, other: &XOnlyPublicKey) -> Ordering

Like cmp::Cmp but faster and with no guarantees across library versions.

The Cmp implementation for FFI types is stable but slow because it first serializes self and other before comparing them. This function provides a faster comparison if you know that your types come from the same library version.

pub fn eq_fast_unstable(&self, other: &XOnlyPublicKey) -> bool

Like cmp::Eq but faster and with no guarantees across library versions.

The Eq implementation for FFI types is stable but slow because it first serializes self and other before comparing them. This function provides a faster equality check if you know that your types come from the same library version.

impl XOnlyPublicKey

pub fn as_ptr(&self) -> *const XOnlyPublicKey

๐Deprecated since 0.25.0: Use Self::as_c_ptr if you need to access the FFI layer

Obtains a raw const pointer suitable for use with FFI functions.

pub fn as_mut_ptr(&mut self) -> *mut XOnlyPublicKey

๐Deprecated since 0.25.0: Use Self::as_mut_c_ptr if you need to access the FFI layer

Obtains a raw mutable pointer suitable for use with FFI functions.

pub fn from_keypair(keypair: &KeyPair) -> (XOnlyPublicKey, Parity)

Returns the XOnlyPublicKey (and itโs Parity) for keypair.

pub fn from_slice(data: &[u8]) -> Result<XOnlyPublicKey, Error>

Creates a schnorr public key directly from a slice.

Errors

Returns Error::InvalidPublicKey if the length of the data slice is not 32 bytes or the slice does not represent a valid Secp256k1 point x coordinate.

pub fn serialize(&self) -> [u8; 32]

Serializes the key as a byte-encoded x coordinate value (32 bytes).

pub fn add_tweak<V>( self, secp: &Secp256k1<V>, tweak: &Scalar ) -> Result<(XOnlyPublicKey, Parity), Error>where V: Verification,

Tweaks an XOnlyPublicKey by adding the generator multiplied with the given tweak to it.

Returns

The newly tweaked key plus an opaque type representing the parity of the tweaked key, this should be provided to tweak_add_check which can be used to verify a tweak more efficiently than regenerating it and checking equality.

Errors

If the resulting key would be invalid.

Examples

use secp256k1::{Secp256k1, KeyPair, Scalar, XOnlyPublicKey};

let secp = Secp256k1::new();
let tweak = Scalar::random();

let mut key_pair = KeyPair::new(&secp, &mut rand::thread_rng());
let (xonly, _parity) = key_pair.x_only_public_key();
let tweaked = xonly.add_tweak(&secp, &tweak).expect("Improbable to fail with a randomly generated tweak");

pub fn tweak_add_check<V>( &self, secp: &Secp256k1<V>, tweaked_key: &XOnlyPublicKey, tweaked_parity: Parity, tweak: Scalar ) -> boolwhere V: Verification,

Verifies that a tweak produced by XOnlyPublicKey::add_tweak was computed correctly.

Should be called on the original untweaked key. Takes the tweaked key and output parity from XOnlyPublicKey::add_tweak as input.

Currently this is not much more efficient than just recomputing the tweak and checking equality. However, in future this API will support batch verification, which is significantly faster, so it is wise to design protocols with this in mind.

Returns

True if tweak and check is successful, false otherwise.

Examples

use secp256k1::{Secp256k1, KeyPair, Scalar};

let secp = Secp256k1::new();
let tweak = Scalar::random();

let mut key_pair = KeyPair::new(&secp, &mut rand::thread_rng());
let (mut public_key, _) = key_pair.x_only_public_key();
let original = public_key;
let (tweaked, parity) = public_key.add_tweak(&secp, &tweak).expect("Improbable to fail with a randomly generated tweak");
assert!(original.tweak_add_check(&secp, &tweaked, parity, tweak));

pub fn public_key(&self, parity: Parity) -> PublicKey

Returns the PublicKey for this XOnlyPublicKey.

This is equivalent to using [PublicKey::from_xonly_and_parity(self, parity)].