Trait bitcoin::key::secp256k1::Context

pub unsafe trait Context: Sealed {
    const FLAGS: u32;
    const DESCRIPTION: &'static str;

    // Required method
    unsafe fn deallocate(ptr: *mut u8, size: usize);
}

A trait for all kinds of contexts that lets you define the exact flags and a function to deallocate memory. It isn’t possible to implement this for types outside this crate.

Safety

This trait is marked unsafe to allow unsafe implementations of deallocate.

Required Associated Constants

const FLAGS: u32

Flags for the ffi.

const DESCRIPTION: &'static str

A constant description of the context.

Required Methods

unsafe fn deallocate(ptr: *mut u8, size: usize)

A function to deallocate the memory when the context is dropped.

Safety

ptr must be valid. Further safety constraints may be imposed by std::alloc::dealloc.