Enum bitcoin::network::message::NetworkMessage

pub enum NetworkMessage {

    Version(VersionMessage),
    Verack,
    Addr(Vec<(u32, Address)>),
    Inv(Vec<Inventory>),
    GetData(Vec<Inventory>),
    NotFound(Vec<Inventory>),
    GetBlocks(GetBlocksMessage),
    GetHeaders(GetHeadersMessage),
    MemPool,
    Tx(Transaction),
    Block(Block),
    Headers(Vec<BlockHeader>),
    SendHeaders,
    GetAddr,
    Ping(u64),
    Pong(u64),
    GetCFilters(GetCFilters),
    CFilter(CFilter),
    GetCFHeaders(GetCFHeaders),
    CFHeaders(CFHeaders),
    GetCFCheckpt(GetCFCheckpt),
    CFCheckpt(CFCheckpt),
    Alert(Vec<u8>),
    Reject(Reject),
    FeeFilter(i64),
    WtxidRelay,
    AddrV2(Vec<AddrV2Message>),
    SendAddrV2,
    Unknown {
        command: CommandString,
        payload: Vec<u8>,
    },

}

A Network message payload. Proper documentation is available on at Bitcoin Wiki: Protocol Specification

Variants

Version(VersionMessage)

Tuple Fields

0: VersionMessage

version

Verack

verack

Addr(Vec<(u32, Address)>)

Tuple Fields

0: Vec<(u32, Address)>

addr

Inv(Vec<Inventory>)

Tuple Fields

0: Vec<Inventory>

inv

GetData(Vec<Inventory>)

Tuple Fields

0: Vec<Inventory>

getdata

NotFound(Vec<Inventory>)

Tuple Fields

0: Vec<Inventory>

notfound

GetBlocks(GetBlocksMessage)

Tuple Fields

0: GetBlocksMessage

getblocks

GetHeaders(GetHeadersMessage)

Tuple Fields

0: GetHeadersMessage

getheaders

MemPool

mempool

Tx(Transaction)

Tuple Fields

0: Transaction

tx

Block(Block)

Tuple Fields

0: Block

block

Headers(Vec<BlockHeader>)

Tuple Fields

0: Vec<BlockHeader>

headers

SendHeaders

sendheaders

GetAddr

getaddr

Ping(u64)

Tuple Fields

0: u64

ping

Pong(u64)

Tuple Fields

0: u64

pong

GetCFilters(GetCFilters)

Tuple Fields

0: GetCFilters

BIP157 getcfilters

CFilter(CFilter)

Tuple Fields

0: CFilter

BIP157 cfilter

GetCFHeaders(GetCFHeaders)

Tuple Fields

0: GetCFHeaders

BIP157 getcfheaders

CFHeaders(CFHeaders)

Tuple Fields

0: CFHeaders

BIP157 cfheaders

GetCFCheckpt(GetCFCheckpt)

Tuple Fields

0: GetCFCheckpt

BIP157 getcfcheckpt

CFCheckpt(CFCheckpt)

Tuple Fields

0: CFCheckpt

BIP157 cfcheckpt

Alert(Vec<u8>)

Tuple Fields

0: Vec<u8>

alert

Reject(Reject)

Tuple Fields

0: Reject

reject

FeeFilter(i64)

Tuple Fields

0: i64

feefilter

WtxidRelay

wtxidrelay

AddrV2(Vec<AddrV2Message>)

Tuple Fields

0: Vec<AddrV2Message>

addrv2

SendAddrV2

sendaddrv2

Unknown

Fields

command: CommandString

The command of this message.

payload: Vec<u8>

The payload of this message.

Any other message.

Implementations

impl NetworkMessage

pub fn cmd(&self) -> &'static str

Return the message command as a static string reference.

This returns "unknown" for NetworkMessage::Unknown, regardless of the actual command in the unknown message. Use the Self::command method to get the command for unknown messages.

pub fn command(&self) -> CommandString

Return the CommandString for the message command.