bitcoin::pow

Struct Target

pub struct Target(/* private fields */);

A 256 bit integer representing target.

The SHA-256 hash of a block’s header must be lower than or equal to the current target for the block to be accepted by the network. The lower the target, the more difficult it is to generate a block. (See also Work.)

ref: https://en.bitcoin.it/wiki/Target

Implementations

impl Target

pub const ZERO: Target

When parsing nBits, Bitcoin Core converts a negative target threshold into a target of zero.

pub const MAX: Self

The maximum possible target.

This value is used to calculate difficulty, which is defined as how difficult the current target makes it to find a block relative to how difficult it would be at the highest possible target. Remember highest target == lowest difficulty.

ref: https://en.bitcoin.it/wiki/Target

pub const fn max_value() -> Self

The maximum possible target (see Target::MAX).

This is provided for consistency with Rust 1.41.1, newer code should use Target::MAX.

pub fn from_compact(c: CompactTarget) -> Target

Computes the Target value from a compact representation.

ref: https://developer.bitcoin.org/reference/block_chain.html#target-nbits

pub fn to_compact_lossy(self) -> CompactTarget

Computes the compact value from a Target representation.

The compact form is by definition lossy, this means that t == Target::from_compact(t.to_compact_lossy()) does not always hold.

pub fn is_met_by(&self, hash: BlockHash) -> bool

Returns true if block hash is less than or equal to this Target.

Proof-of-work validity for a block requires the hash of the block to be less than or equal to the target.

pub fn to_work(self) -> Work

Converts this Target to Work.

“Work” is defined as the work done to mine a block with this target value (recorded in the block header in compact form as nBits). This is not the same as the difficulty to mine a block with this target (see Self::difficulty).

pub fn difficulty(&self) -> u128

Computes the popular “difficulty” measure for mining.

Difficulty represents how difficult the current target makes it to find a block, relative to how difficult it would be at the highest possible target (highest target == lowest difficulty).

For example, a difficulty of 6,695,826 means that at a given hash rate, it will, on average, take ~6.6 million times as long to find a valid block as it would at a difficulty of 1, or alternatively, it will take, again on average, ~6.6 million times as many hashes to find a valid block

Note

Difficulty is calculated using the following algorithm max / current where max is defined for the Bitcoin network and current is the current target for this block. As such, a low target implies a high difficulty. Since Target is represented as a 256 bit integer but difficulty() returns only 128 bits this means for targets below approximately 0xffff_ffff_ffff_ffff_ffff_ffff difficulty() will saturate at u128::MAX.

pub fn difficulty_float(&self) -> f64

Computes the popular “difficulty” measure for mining and returns a float value of f64.

See difficulty for details.

pub fn min_difficulty_transition_threshold(&self) -> Self

Computes the minimum valid Target threshold allowed for a block in which a difficulty adjustment occurs.

The difficulty can only decrease or increase by a factor of 4 max on each difficulty adjustment period.

pub fn max_difficulty_transition_threshold(&self) -> Self

Computes the maximum valid Target threshold allowed for a block in which a difficulty adjustment occurs.

The difficulty can only decrease or increase by a factor of 4 max on each difficulty adjustment period.

impl Target

pub fn from_be_bytes(bytes: [u8; 32]) -> Target

Creates Self from a big-endian byte array.

pub fn from_le_bytes(bytes: [u8; 32]) -> Target

Creates Self from a little-endian byte array.

pub fn to_be_bytes(self) -> [u8; 32]

Converts self to a big-endian byte array.

pub fn to_le_bytes(self) -> [u8; 32]

Converts self to a little-endian byte array.