bitcoin::psbt

Type Alias Psbt

pub type Psbt = PartiallySignedTransaction;

Partially signed transaction, commonly referred to as a PSBT.

Aliased Type

pub struct Psbt {
    pub unsigned_tx: Transaction,
    pub version: u32,
    pub xpub: BTreeMap<ExtendedPubKey, (Fingerprint, DerivationPath)>,
    pub proprietary: BTreeMap<ProprietaryKey, Vec<u8>>,
    pub unknown: BTreeMap<Key, Vec<u8>>,
    pub inputs: Vec<Input>,
    pub outputs: Vec<Output>,
}

Fields
unsigned_tx: Transaction

The unsigned transaction, scriptSigs and witnesses for each input must be empty.

version: u32

The version number of this PSBT. If omitted, the version number is 0.

xpub: BTreeMap<ExtendedPubKey, (Fingerprint, DerivationPath)>

A global map from extended public keys to the used key fingerprint and derivation path as defined by BIP 32.

proprietary: BTreeMap<ProprietaryKey, Vec<u8>>

Global proprietary key-value pairs.

unknown: BTreeMap<Key, Vec<u8>>

Unknown global key-value pairs.

inputs: Vec<Input>

The corresponding key-value map for each input in the unsigned transaction.

outputs: Vec<Output>

The corresponding key-value map for each output in the unsigned transaction.