Module bitcoin::sighash

Signature hash implementation (used in transaction signing).

Efficient implementation of the algorithm to compute the message to be signed according to Bip341, Bip143 and legacy (before Bip143).

Computing signature hashes is required to sign a transaction and this module is designed to handle its complexity efficiently. Computing these hashes is as simple as creating SighashCache and calling its methods.

Structs

Annex

The Annex struct is a slice wrapper enforcing first byte is 0x50.

LegacySighash

Hash of a transaction according to the legacy signature algorithm.

NonStandardSighashType

This type is consensus valid but an input including it would prevent the transaction from being relayed on today’s Bitcoin network.

ScriptPath

Information related to the script path spending.

SegwitV0Sighash

Hash of a transaction according to the segwit version 0 signature algorithm.

SighashCache

Efficiently calculates signature hash message for legacy, segwit and taproot inputs.

SighashTypeParseError

Error returned for failure during parsing one of the sighash types.

TapSighash

Taproot-tagged hash with tag “TapSighash”.

TapSighashTag

The tag used for TapSighash

Enums

EcdsaSighashType

Hashtype of an input’s signature, encoded in the last byte of the signature.

Error

Possible errors in computing the signature message.

Prevouts

Contains outputs of previous transactions. In the case TapSighashType variant is SIGHASH_ANYONECANPAY, Prevouts::One may be used.

TapSighashType

Hashtype of an input’s signature, encoded in the last byte of the signature. Fixed values so they can be cast as integer types for encoding.