Struct bitcoin::ScriptBuf

pub struct ScriptBuf(_);

An owned, growable script.

ScriptBuf is the most common script type that has the ownership over the contents of the script. It has a close relationship with its borrowed counterpart, Script.

Just as other similar types, this implements Deref, so deref coercions apply. Also note that all the safety/validity restrictions that apply to Script apply to ScriptBuf as well.

Implementations

impl ScriptBuf

pub fn new() -> Self

Creates a new empty script.

pub fn with_capacity(capacity: usize) -> Self

Creates a new empty script with pre-allocated capacity.

pub fn reserve(&mut self, additional_len: usize)

Pre-allocates at least additional_len bytes if needed.

Reserves capacity for at least additional_len more bytes to be inserted in the given script. The script may reserve more space to speculatively avoid frequent reallocations. After calling reserve, capacity will be greater than or equal to self.len() + additional_len. Does nothing if capacity is already sufficient.

Panics

Panics if the new capacity exceeds isize::MAX bytes.

pub fn reserve_exact(&mut self, additional_len: usize)

Pre-allocates exactly additional_len bytes if needed.

Unlike reserve, this will not deliberately over-allocate to speculatively avoid frequent allocations. After calling reserve_exact, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore, capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Panics

Panics if the new capacity exceeds isize::MAX bytes.

pub fn as_script(&self) -> &Script

Returns a reference to unsized script.

pub fn as_mut_script(&mut self) -> &mut Script

Returns a mutable reference to unsized script.

pub fn builder() -> Builder

Creates a new script builder

pub fn new_p2pk(pubkey: &PublicKey) -> Self

Generates P2PK-type of scriptPubkey.

pub fn new_p2pkh(pubkey_hash: &PubkeyHash) -> Self

Generates P2PKH-type of scriptPubkey.

pub fn new_p2sh(script_hash: &ScriptHash) -> Self

Generates P2SH-type of scriptPubkey with a given hash of the redeem script.

pub fn new_v0_p2wpkh(pubkey_hash: &WPubkeyHash) -> Self

Generates P2WPKH-type of scriptPubkey.

pub fn new_v0_p2wsh(script_hash: &WScriptHash) -> Self

Generates P2WSH-type of scriptPubkey with a given hash of the redeem script.

pub fn new_v1_p2tr<C: Verification>( secp: &Secp256k1<C>, internal_key: UntweakedPublicKey, merkle_root: Option<TapNodeHash> ) -> Self

Generates P2TR for script spending path using an internal public key and some optional script tree merkle root.

pub fn new_v1_p2tr_tweaked(output_key: TweakedPublicKey) -> Self

Generates P2TR for key spending path for a known TweakedPublicKey.

pub fn new_witness_program(witness_program: &WitnessProgram) -> Self

Generates P2WSH-type of scriptPubkey with a given WitnessProgram.

pub fn new_op_return<T: AsRef<PushBytes>>(data: &T) -> Self

Generates OP_RETURN-type of scriptPubkey for the given data.

pub fn from_hex(s: &str) -> Result<Self, Error>

Creates a ScriptBuf from a hex string.

pub fn from_bytes(bytes: Vec<u8>) -> Self

Converts byte vector into script.

This method doesn’t (re)allocate.

pub fn into_bytes(self) -> Vec<u8>

Converts the script into a byte vector.

This method doesn’t (re)allocate.

pub fn to_p2sh(&self) -> ScriptBuf

Computes the P2SH output corresponding to this redeem script.

pub fn p2wpkh_script_code(&self) -> Option<ScriptBuf>

Returns the script code used for spending a P2WPKH output if this script is a script pubkey for a P2WPKH output. The scriptCode is described in BIP143.

pub fn push_opcode(&mut self, data: All)

Adds a single opcode to the script.

pub fn push_slice<T: AsRef<PushBytes>>(&mut self, data: T)

Adds instructions to push some arbitrary data onto the stack.

pub fn push_instruction(&mut self, instruction: Instruction<'_>)

Add a single instruction to the script.

Panics

The method panics if the instruction is a data push with length greater or equal to 0x100000000.

pub fn push_instruction_no_opt(&mut self, instruction: Instruction<'_>)

Like push_instruction, but avoids calling reserve to not re-check the length.

pub fn scan_and_push_verify(&mut self)

Adds an OP_VERIFY to the script or replaces the last opcode with VERIFY form.

Some opcodes such as OP_CHECKSIG have a verify variant that works as if VERIFY was in the script right after. To save space this function appends VERIFY only if the most-recently-added opcode does not have an alternate VERIFY form. If it does the last opcode is replaced. E.g., OP_CHECKSIG will become OP_CHECKSIGVERIFY.

Note that existing OP_*VERIFY opcodes do not lead to the instruction being ignored because OP_VERIFY consumes an item from the stack so ignoring them would change the semantics.

This function needs to iterate over the script to find the last instruction. Prefer Builder if you’re creating the script from scratch or if you want to push OP_VERIFY multiple times.

pub fn into_boxed_script(self) -> Box<Script>

Converts this ScriptBuf into a boxed Script.

This method reallocates if the capacity is greater than lenght of the script but should not when they are equal. If you know beforehand that you need to create a script of exact size use reserve_exact before adding data to the script so that the reallocation can be avoided.