Struct bitcoin::SignedAmount

pub struct SignedAmount(_);

SignedAmount

The SignedAmount type can be used to express Bitcoin amounts that supports arithmetic and conversion to various denominations.

Warning!

This type implements several arithmetic operations from core::ops. To prevent errors due to overflow or underflow when using these operations, it is advised to instead use the checked arithmetic methods whose names start with checked_. The operations from core::ops that Amount implements will panic when overflow or underflow occurs.

Implementations

impl SignedAmount

pub const ZERO: SignedAmount

The zero amount.

pub const ONE_SAT: SignedAmount

Exactly one satoshi.

pub const ONE_BTC: SignedAmount

Exactly one bitcoin.

pub fn from_sat(satoshi: i64) -> SignedAmount

Create an SignedAmount with satoshi precision and the given number of satoshis.

pub fn as_sat(self) -> i64

Get the number of satoshis in this SignedAmount.

pub fn max_value() -> SignedAmount

The maximum value of an SignedAmount.

pub fn min_value() -> SignedAmount

The minimum value of an SignedAmount.

pub fn from_btc(btc: f64) -> Result<SignedAmount, ParseAmountError>

Convert from a value expressing bitcoins to an SignedAmount.

pub fn from_str_in(
s: &str,
denom: Denomination
) -> Result<SignedAmount, ParseAmountError>

Parse a decimal string as a value in the given denomination.

Note: This only parses the value string. If you want to parse a value with denomination, use FromStr.

pub fn from_str_with_denomination(
s: &str
) -> Result<SignedAmount, ParseAmountError>

Parses amounts with denomination suffix like they are produced with Self::to_string_with_denomination or with fmt::Display. If you want to parse only the amount without the denomination, use Self::from_str_in.

pub fn to_float_in(self, denom: Denomination) -> f64

Express this SignedAmount as a floating-point value in the given denomination.

Please be aware of the risk of using floating-point numbers.

pub fn as_btc(self) -> f64

Express this SignedAmount as a floating-point value in Bitcoin.

Equivalent to to_float_in(Denomination::Bitcoin).

Please be aware of the risk of using floating-point numbers.

pub fn from_float_in(
value: f64,
denom: Denomination
) -> Result<SignedAmount, ParseAmountError>

Convert this SignedAmount in floating-point notation with a given denomination. Can return error if the amount is too big, too precise or negative.

Please be aware of the risk of using floating-point numbers.

pub fn fmt_value_in(
self,
f: &mut dyn Write,
denom: Denomination
) -> Result<(), Error>

Format the value of this SignedAmount in the given denomination.

Does not include the denomination.

pub fn to_string_in(self, denom: Denomination) -> String

Get a string number of this SignedAmount in the given denomination.

Does not include the denomination.

pub fn to_string_with_denomination(self, denom: Denomination) -> String

Get a formatted string of this SignedAmount in the given denomination, suffixed with the abbreviation for the denomination.

pub fn abs(self) -> SignedAmount

Get the absolute value of this SignedAmount.

pub fn signum(self) -> i64

Returns a number representing sign of this SignedAmount.

0 if the amount is zero

1 if the amount is positive

-1 if the amount is negative

pub fn is_positive(self) -> bool

Returns true if this SignedAmount is positive and false if this SignedAmount is zero or negative.

pub fn is_negative(self) -> bool

Returns true if this SignedAmount is negative and false if this SignedAmount is zero or positive.

pub fn checked_abs(self) -> Option<SignedAmount>

Get the absolute value of this SignedAmount. Returns None if overflow occurred. (self == min_value())

pub fn checked_add(self, rhs: SignedAmount) -> Option<SignedAmount>

Checked addition. Returns None if overflow occurred.

pub fn checked_sub(self, rhs: SignedAmount) -> Option<SignedAmount>

Checked subtraction. Returns None if overflow occurred.

pub fn checked_mul(self, rhs: i64) -> Option<SignedAmount>

Checked multiplication. Returns None if overflow occurred.

pub fn checked_div(self, rhs: i64) -> Option<SignedAmount>

Checked integer division. Be aware that integer division loses the remainder if no exact division can be made. Returns None if overflow occurred.

pub fn checked_rem(self, rhs: i64) -> Option<SignedAmount>

Checked remainder. Returns None if overflow occurred.

pub fn positive_sub(self, rhs: SignedAmount) -> Option<SignedAmount>

Subtraction that doesn’t allow negative SignedAmounts. Returns None if either [self], rhs or the result is strictly negative.

pub fn to_unsigned(self) -> Result<Amount, ParseAmountError>

Convert to an unsigned amount.