Struct bitcoin::Transaction

pub struct Transaction {
    pub version: i32,
    pub lock_time: u32,
    pub input: Vec<TxIn, Global>,
    pub output: Vec<TxOut, Global>,
}

A Bitcoin transaction, which describes an authenticated movement of coins.

If any inputs have nonempty witnesses, the entire transaction is serialized in the post-BIP141 Segwit format which includes a list of witnesses. If all inputs have empty witnesses, the transaction is serialized in the pre-BIP141 format.

There is one major exception to this: to avoid deserialization ambiguity, if the transaction has no inputs, it is serialized in the BIP141 style. Be aware that this differs from the transaction format in PSBT, which never uses BIP141. (Ordinarily there is no conflict, since in PSBT transactions are always unsigned and therefore their inputs have empty witnesses.)

The specific ambiguity is that Segwit uses the flag bytes 0001 where an old serializer would read the number of transaction inputs. The old serializer would interpret this as “no inputs, one output”, which means the transaction is invalid, and simply reject it. Segwit further specifies that this encoding should only be used when some input has a nonempty witness; that is, witness-less transactions should be encoded in the traditional format.

However, in protocols where transactions may legitimately have 0 inputs, e.g. when parties are cooperatively funding a transaction, the “00 means Segwit” heuristic does not work. Since Segwit requires such a transaction be encoded in the original transaction format (since it has no inputs and therefore no input witnesses), a traditionally encoded transaction may have the 0001 Segwit flag in it, which confuses most Segwit parsers including the one in Bitcoin Core.

We therefore deviate from the spec by always using the Segwit witness encoding for 0-input transactions, which results in unambiguously parseable transactions.

Fields
version: i32

The protocol version, is currently expected to be 1 or 2 (BIP 68).

lock_time: u32

Block number before which this transaction is valid, or 0 for valid immediately.

input: Vec<TxIn, Global>

List of inputs

output: Vec<TxOut, Global>

List of outputs

Implementations

impl Transaction

pub fn ntxid(&self) -> Hash

Computes a “normalized TXID” which does not include any signatures. This gives a way to identify a transaction that is “the same” as another in the sense of having same inputs and outputs.

pub fn txid(&self) -> Txid

Computes the txid. For non-segwit transactions this will be identical to the output of wtxid(), but for segwit transactions, this will give the correct txid (not including witnesses) while wtxid will also hash witnesses.

pub fn wtxid(&self) -> Wtxid

Computes SegWit-version of the transaction id (wtxid). For transaction with the witness data this hash includes witness, for pre-witness transaction it is equal to the normal value returned by txid() function.

pub fn encode_signing_data_to<Write, U>(
&self,
writer: Write,
input_index: usize,
script_pubkey: &Script,
sighash_type: U
) -> Result<(), Error> where
Write: Write,
U: Into<u32>,

Encodes the signing data from which a signature hash for a given input index with a given sighash flag can be computed. To actually produce a scriptSig, this hash needs to be run through an ECDSA signer, the SigHashType appended to the resulting sig, and a script written around this, but this is the general (and hard) part.

The sighash_type supports arbitrary u32 value, instead of just [SigHashType], because internally 4 bytes are being hashed, even though only lowest byte is appended to signature in a transaction.

Warning This does NOT attempt to support OP_CODESEPARATOR. In general this would require evaluating script_pubkey to determine which separators get evaluated and which don’t, which we don’t have the information to determine.

Panics

Panics if input_index is greater than or equal to self.input.len()

pub fn signature_hash(
&self,
input_index: usize,
script_pubkey: &Script,
sighash_u32: u32
) -> SigHash

Computes a signature hash for a given input index with a given sighash flag. To actually produce a scriptSig, this hash needs to be run through an ECDSA signer, the SigHashType appended to the resulting sig, and a script written around this, but this is the general (and hard) part.

Warning This does NOT attempt to support OP_CODESEPARATOR. In general this would require evaluating script_pubkey to determine which separators get evaluated and which don’t, which we don’t have the information to determine.

Panics

Panics if input_index is greater than or equal to self.input.len()

pub fn get_weight(&self) -> usize

Gets the “weight” of this transaction, as defined by BIP141. For transactions with an empty witness, this is simply the consensus-serialized size times 4. For transactions with a witness, this is the non-witness consensus-serialized size multiplied by 3 plus the with-witness consensus-serialized size.

pub fn get_size(&self) -> usize

Gets the regular byte-wise consensus-serialized size of this transaction.

pub fn is_coin_base(&self) -> bool

Is this a coin base transaction?

pub fn is_explicitly_rbf(&self) -> bool

Returns true if the transaction itself opted in to be BIP-125-replaceable (RBF). This does not cover the case where a transaction becomes replaceable due to ancestors being RBF.