bitcoin::taproot

Struct TapNodeHash

pub struct TapNodeHash(/* private fields */);

Tagged hash used in taproot trees; see BIP-340 for tagging rules

Implementations

impl TapNodeHash

pub fn from_raw_hash(inner: Hash<TapBranchTag>) -> TapNodeHash

Creates this wrapper type from the inner hash type.

pub fn to_raw_hash(self) -> Hash<TapBranchTag>

Returns the inner hash (sha256, sh256d etc.).

pub fn as_raw_hash(&self) -> &Hash<TapBranchTag>

Returns a reference to the inner hash (sha256, sh256d etc.).

impl TapNodeHash

pub fn from_node_hashes(a: TapNodeHash, b: TapNodeHash) -> TapNodeHash

Computes branch hash given two hashes of the nodes underneath it.

pub fn assume_hidden(hash: [u8; 32]) -> TapNodeHash

Assumes the given 32 byte array as hidden TapNodeHash.

Similar to TapLeafHash::from_byte_array, but explicitly conveys that the hash is constructed from a hidden node. This also has better ergonomics because it does not require the caller to import the Hash trait.

pub fn from_script(script: &Script, ver: LeafVersion) -> TapNodeHash

Computes the TapNodeHash from a script and a leaf version.