Struct bitcoin::taproot::TaprootMerkleBranch

pub struct TaprootMerkleBranch(_);

The merkle proof for inclusion of a tree in a taptree hash.

Implementations

impl TaprootMerkleBranch

pub fn as_inner(&self) -> &[TapNodeHash]

Returns a reference to the inner vector of hashes.

pub fn len(&self) -> usize

Returns the number of nodes in this merkle proof.

pub fn is_empty(&self) -> bool

Checks if this merkle proof is empty.

pub fn decode(sl: &[u8]) -> Result<Self, TaprootError>

Decodes bytes from control block.

This reads the branch as encoded in the control block: the concatenated 32B byte chunks - one for each hash.

Errors

The function returns an error if the the number of bytes is not an integer multiple of 32 or if the number of hashes exceeds 128.

pub fn encode<Write: Write>(&self, writer: Write) -> Result<usize>

Serializes to a writer.

Returns

The number of bytes written to the writer.

pub fn serialize(&self) -> Vec<u8>

Serializes self as bytes.

pub fn into_inner(self) -> Vec<TapNodeHash>

Returns the inner list of hashes.