Module bitcoin::types

Re-exports

pub use crate::merkle::MerkleProof;

Structs

Block

Bitcoin block: header and transactions

BlockBuilder

Generates a new block mined with the given difficulty

BlockChain

Representation of a Bitcoin blockchain

BlockHeader

Structs Bitcoin Basic Block Headers

FullTransactionProof

We also check the coinbase proof in order to defend against the ‘leaf-node weakness’. See https://bitslog.com/2018/06/09/leaf-node-weakness-in-bitcoin-merkle-tree-design/ .

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H256Le

Represents a bitcoin 32 bytes hash digest encoded in little-endian

Transaction

Bitcoin transaction

TransactionBuilder

Construct txs from inputs and outputs

TransactionInput

Bitcoin transaction input

TransactionInputBuilder

Construct transaction inputs

TransactionOutput

Bitcoin transaction output

U256

Little-endian large integer type 256-bit unsigned integer.

Enums

LockTime

OpCode

Bitcoin Script OpCodes https://github.com/bitcoin/bitcoin/blob/master/src/script/script.h

TransactionInputSource

Constants

HASH160_SIZE_HEX

HASH256_SIZE_HEX

MAX_OPRETURN_SIZE

P2PKH_SCRIPT_SIZE

Custom Types

P2SH_SCRIPT_SIZE

P2WPKH_V0_SCRIPT_SIZE

P2WSH_V0_SCRIPT_SIZE

Type Definitions

Value