Module bitcoin::util

Utility functions.

Functions needed by all parts of the Bitcoin library.

Modules

address

Bitcoin addresses.

amount

Bitcoin amounts.

base58

Base58 encoder and decoder.

bip32

BIP32 implementation.

bip143

BIP143 implementation.

bip152

BIP152 Compact Blocks

bip158

BIP158 Compact Block Filters for light clients.

ecdsa

ECDSA Bitcoin signatures.

hash

Bitcoin hash functions.

key

Bitcoin keys.

merkleblock

Merkle Block and Partial Merkle Tree.

misc

Miscellaneous functions.

psbt

Partially Signed Bitcoin Transactions.

schnorr

Schnorr Bitcoin keys.

sighash

Generalized, efficient, signature hash implementation.

taproot

Bitcoin Taproot.

uint

Big unsigned integer types.

Enums

Error

A general error code, other errors should implement conversions to/from this if appropriate.

Traits

BitArray

A trait which allows numbers to act as fixed-size bit arrays