Enum bitcoin::util::address::WitnessVersion

#[repr(u8)]
pub enum WitnessVersion {

    V0,
    V1,
    V2,
    V3,
    V4,
    V5,
    V6,
    V7,
    V8,
    V9,
    V10,
    V11,
    V12,
    V13,
    V14,
    V15,
    V16,

}

Version of the witness program.

Helps limit possible versions of the witness according to the specification. If a plain u8 type was used instead it would mean that the version may be > 16, which would be incorrect.

First byte of scriptPubkey in transaction output for transactions starting with opcodes ranging from 0 to 16 (inclusive).

Variants

V0

Initial version of witness program. Used for P2WPKH and P2WPK outputs

V1

Version of witness program used for Taproot P2TR outputs.

V2

Future (unsupported) version of witness program.

V3

Future (unsupported) version of witness program.

V4

Future (unsupported) version of witness program.

V5

Future (unsupported) version of witness program.

V6

Future (unsupported) version of witness program.

V7

Future (unsupported) version of witness program.

V8

Future (unsupported) version of witness program.

V9

Future (unsupported) version of witness program.

V10

Future (unsupported) version of witness program.

V11

Future (unsupported) version of witness program.

V12

Future (unsupported) version of witness program.

V13

Future (unsupported) version of witness program.

V14

Future (unsupported) version of witness program.

V15

Future (unsupported) version of witness program.

V16

Future (unsupported) version of witness program.

Implementations

impl WitnessVersion

pub fn from_u5(value: u5) -> Result<Self, Error>

👎Deprecated since 0.29.0:
use try_from instead

Converts 5-bit unsigned integer value matching single symbol from Bech32(m) address encoding (bech32::u5) into WitnessVersion variant.

Returns

Version of the Witness program.

Errors

If the integer does not correspond to any witness version, errors with Error::InvalidWitnessVersion.

pub fn from_num(no: u8) -> Result<Self, Error>

👎Deprecated since 0.29.0:
use try_from instead

Converts an 8-bit unsigned integer value into WitnessVersion variant.

Returns

Version of the Witness program.

Errors

If the integer does not correspond to any witness version, errors with Error::InvalidWitnessVersion.

pub fn from_opcode(opcode: All) -> Result<Self, Error>

👎Deprecated since 0.29.0:
use try_from instead

Converts bitcoin script opcode into WitnessVersion variant.

Returns

Version of the Witness program (for opcodes in range of OP_0..OP_16).

Errors

If the opcode does not correspond to any witness version, errors with Error::MalformedWitnessVersion.

pub fn from_instruction(instruction: Instruction<'_>) -> Result<Self, Error>

👎Deprecated since 0.29.0:
use try_from instead

Converts bitcoin script Instruction (parsed opcode) into WitnessVersion variant.

Returns

Version of the Witness program for Instruction::Op and Instruction::PushBytes with byte value within 1..=16 range.

Errors

If the opcode does not correspond to any witness version, errors with Error::MalformedWitnessVersion for the rest of opcodes.

pub fn into_num(self) -> u8

👎Deprecated since 0.29.0:
use to_num instead

Returns integer version number representation for a given WitnessVersion value.

NB: this is not the same as an integer representation of the opcode signifying witness version in bitcoin script. Thus, there is no function to directly convert witness version into a byte since the conversion requires context (bitcoin script or just a version number).

pub fn to_num(self) -> u8

Returns integer version number representation for a given WitnessVersion value.

NB: this is not the same as an integer representation of the opcode signifying witness version in bitcoin script. Thus, there is no function to directly convert witness version into a byte since the conversion requires context (bitcoin script or just a version number).

pub fn bech32_variant(&self) -> Variant

Determines the checksum variant. See BIP-0350 for specification.