Struct bitcoin::util::address::Address

pub struct Address {
    pub payload: Payload,
    pub network: Network,
}

A Bitcoin address.

Relevant BIPs

BIP13 - Address Format for pay-to-script-hash

BIP16 - Pay to Script Hash

BIP141 - Segregated Witness (Consensus layer)

BIP142 - Address Format for Segregated Witness

BIP341 - Taproot: SegWit version 1 spending rules

BIP350 - Bech32m format for v1+ witness addresses

Fields
payload: Payload

The type of the address.

network: Network

The network on which this address is usable.

Implementations

impl Address

pub fn p2pkh(pk: &PublicKey, network: Network) -> Address

Creates a pay to (compressed) public key hash address from a public key.

This is the preferred non-witness type address.

pub fn p2sh(script: &Script, network: Network) -> Result<Address, Error>

Creates a pay to script hash P2SH address from a script.

This address type was introduced with BIP16 and is the popular type to implement multi-sig these days.

pub fn p2wpkh(pk: &PublicKey, network: Network) -> Result<Address, Error>

Creates a witness pay to public key address from a public key.

This is the native segwit address type for an output redeemable with a single signature.

Errors

Will only return an error if an uncompressed public key is provided.

pub fn p2shwpkh(pk: &PublicKey, network: Network) -> Result<Address, Error>

Creates a pay to script address that embeds a witness pay to public key.

This is a segwit address type that looks familiar (as p2sh) to legacy clients.

Errors

Will only return an Error if an uncompressed public key is provided.

pub fn p2wsh(script: &Script, network: Network) -> Address

Creates a witness pay to script hash address.

pub fn p2shwsh(script: &Script, network: Network) -> Address

Creates a pay to script address that embeds a witness pay to script hash address.

This is a segwit address type that looks familiar (as p2sh) to legacy clients.

pub fn p2tr<C: Verification>(
secp: &Secp256k1<C>,
internal_key: UntweakedPublicKey,
merkle_root: Option<TapBranchHash>,
network: Network
) -> Address

Creates a pay to taproot address from an untweaked key.

pub fn p2tr_tweaked(output_key: TweakedPublicKey, network: Network) -> Address

Creates a pay to taproot address from a pre-tweaked output key.

This method is not recommended for use, Address::p2tr() should be used where possible.

pub fn address_type(&self) -> Option<AddressType>

Gets the address type of the address.

Returns

None if unknown, non-standard or related to the future witness version.

pub fn is_standard(&self) -> bool

Checks whether or not the address is following Bitcoin standardness rules.

SegWit addresses with unassigned witness versions or non-standard program sizes are considered non-standard.

pub fn from_script(script: &Script, network: Network) -> Result<Address, Error>

Constructs an Address from an output script (scriptPubkey).

pub fn script_pubkey(&self) -> Script

Generates a script pubkey spending to this address.

pub fn to_qr_uri(&self) -> String

Creates a URI string bitcoin:address optimized to be encoded in QR codes.

If the address is bech32, both the schema and the address become uppercase. If the address is base58, the schema is lowercase and the address is left mixed case.

Quoting BIP 173 “inside QR codes uppercase SHOULD be used, as those permit the use of alphanumeric mode, which is 45% more compact than the normal byte mode.”

pub fn is_valid_for_network(&self, network: Network) -> bool

Parsed addresses do not always have one network. The problem is that legacy testnet, regtest and signet addresse use the same prefix instead of multiple different ones. When parsing, such addresses are always assumed to be testnet addresses (the same is true for bech32 signet addresses). So if one wants to check if an address belongs to a certain network a simple comparison is not enough anymore. Instead this function can be used.

use bitcoin::{Address, Network};

let address: Address = "2N83imGV3gPwBzKJQvWJ7cRUY2SpUyU6A5e".parse().unwrap();
assert!(address.is_valid_for_network(Network::Testnet));
assert!(address.is_valid_for_network(Network::Regtest));
assert!(address.is_valid_for_network(Network::Signet));

assert_eq!(address.is_valid_for_network(Network::Bitcoin), false);

let address: Address = "32iVBEu4dxkUQk9dJbZUiBiQdmypcEyJRf".parse().unwrap();
assert!(address.is_valid_for_network(Network::Bitcoin));
assert_eq!(address.is_valid_for_network(Network::Testnet), false);

pub fn is_related_to_pubkey(&self, pubkey: &PublicKey) -> bool

Returns true if the given pubkey is directly related to the address payload.

This is determined by directly comparing the address payload with either the hash of the given public key or the segwit redeem hash generated from the given key. For taproot addresses, the supplied key is assumed to be tweaked

pub fn is_related_to_xonly_pubkey(&self, xonly_pubkey: &XOnlyPublicKey) -> bool

Returns true if the supplied xonly public key can be used to derive the address.

This will only work for Taproot addresses. The Public Key is assumed to have already been tweaked.