Module bitcoin::util::amount

Amounts

This module mainly introduces the Amount and SignedAmount types. We refer to the documentation on the types for more information.

Structs

Amount

Amount

SignedAmount

SignedAmount

Enums

Denomination

A set of denominations in which amounts can be expressed.

ParseAmountError

An error during amount parsing.