Struct bitcoin::util::amount::Amount

pub struct Amount(_);

Amount

The Amount type can be used to express Bitcoin amounts that supports arithmetic and conversion to various denominations.

Warning!

This type implements several arithmetic operations from core::ops. To prevent errors due to overflow or underflow when using these operations, it is advised to instead use the checked arithmetic methods whose names start with checked_. The operations from core::ops that Amount implements will panic when overflow or underflow occurs. Also note that since the internal representation of amounts is unsigned, subtracting below zero is considered an underflow and will cause a panic if you’re not using the checked arithmetic methods.

Implementations

impl Amount

pub const ZERO: Amount

The zero amount.

pub const ONE_SAT: Amount

Exactly one satoshi.

pub const ONE_BTC: Amount

Exactly one bitcoin.

pub fn from_sat(satoshi: u64) -> Amount

Create an Amount with satoshi precision and the given number of satoshis.

pub fn as_sat(self) -> u64

Get the number of satoshis in this Amount.

pub fn max_value() -> Amount

The maximum value of an Amount.

pub fn min_value() -> Amount

The minimum value of an Amount.

pub fn from_btc(btc: f64) -> Result<Amount, ParseAmountError>

Convert from a value expressing bitcoins to an Amount.

pub fn from_str_in(
s: &str,
denom: Denomination
) -> Result<Amount, ParseAmountError>

Parse a decimal string as a value in the given denomination.

Note: This only parses the value string. If you want to parse a value with denomination, use FromStr.

pub fn from_str_with_denomination(s: &str) -> Result<Amount, ParseAmountError>

Parses amounts with denomination suffix like they are produced with Self::to_string_with_denomination or with fmt::Display. If you want to parse only the amount without the denomination, use Self::from_str_in.

pub fn to_float_in(self, denom: Denomination) -> f64

Express this Amount as a floating-point value in the given denomination.

Please be aware of the risk of using floating-point numbers.

pub fn as_btc(self) -> f64

Express this Amount as a floating-point value in Bitcoin.

Equivalent to to_float_in(Denomination::Bitcoin).

Please be aware of the risk of using floating-point numbers.

pub fn from_float_in(
value: f64,
denom: Denomination
) -> Result<Amount, ParseAmountError>

Convert this Amount in floating-point notation with a given denomination. Can return error if the amount is too big, too precise or negative.

Please be aware of the risk of using floating-point numbers.

pub fn fmt_value_in(self, f: &mut dyn Write, denom: Denomination) -> Result

Format the value of this Amount in the given denomination.

Does not include the denomination.

pub fn to_string_in(self, denom: Denomination) -> String

Get a string number of this Amount in the given denomination.

Does not include the denomination.

pub fn to_string_with_denomination(self, denom: Denomination) -> String

Get a formatted string of this Amount in the given denomination, suffixed with the abbreviation for the denomination.

pub fn checked_add(self, rhs: Amount) -> Option<Amount>

Checked addition. Returns None if overflow occurred.

pub fn checked_sub(self, rhs: Amount) -> Option<Amount>

Checked subtraction. Returns None if overflow occurred.

pub fn checked_mul(self, rhs: u64) -> Option<Amount>

Checked multiplication. Returns None if overflow occurred.

pub fn checked_div(self, rhs: u64) -> Option<Amount>

Checked integer division. Be aware that integer division loses the remainder if no exact division can be made. Returns None if overflow occurred.

pub fn checked_rem(self, rhs: u64) -> Option<Amount>

Checked remainder. Returns None if overflow occurred.

pub fn to_signed(self) -> Result<SignedAmount, ParseAmountError>

Convert to a signed amount.