Struct bitcoin::util::bip143::SigHashCache

pub struct SigHashCache<R: Deref<Target = Transaction>> { /* fields omitted */ }

A replacement for SigHashComponents which supports all sighash modes

Implementations

impl<R: Deref<Target = Transaction>> SigHashCache<R>

pub fn new(tx: R) -> Self

Compute the sighash components from an unsigned transaction and auxiliary in a lazy manner when required. For the generated sighashes to be valid, no fields in the transaction may change except for script_sig and witnesses.

pub fn hash_prevouts(&mut self) -> Hash

Calculate hash for prevouts

pub fn hash_sequence(&mut self) -> Hash

Calculate hash for input sequence values

pub fn hash_outputs(&mut self) -> Hash

Calculate hash for outputs

pub fn encode_signing_data_to<Write: Write>(
&mut self,
writer: Write,
input_index: usize,
script_code: &Script,
value: u64,
sighash_type: SigHashType
) -> Result<(), Error>

Encode the BIP143 signing data for any flag type into a given object implementing a std::io::Write trait.

pub fn signature_hash(
&mut self,
input_index: usize,
script_code: &Script,
value: u64,
sighash_type: SigHashType
) -> SigHash

Compute the BIP143 sighash for any flag type. See SighashComponents::sighash_all simpler API for the most common case

impl<R: DerefMut<Target = Transaction>> SigHashCache<R>

pub fn access_witness(&mut self, input_index: usize) -> &mut Vec<Vec<u8>>

When the SigHashCache is initialized with a mutable reference to a transaction instead of a regular reference, this method is available to allow modification to the witnesses.

This allows in-line signing such as

use bitcoin::blockdata::transaction::{Transaction, SigHashType};
use bitcoin::util::bip143::SigHashCache;
use bitcoin::Script;

let mut tx_to_sign = Transaction { version: 2, lock_time: 0, input: Vec::new(), output: Vec::new() };
let input_count = tx_to_sign.input.len();

let mut sig_hasher = SigHashCache::new(&mut tx_to_sign);
for inp in 0..input_count {
    let prevout_script = Script::new();
    let _sighash = sig_hasher.signature_hash(inp, &prevout_script, 42, SigHashType::All);
    // ... sign the sighash
    sig_hasher.access_witness(inp).push(Vec::new());
}