Struct bitcoin::util::bip32::ExtendedPubKey

pub struct ExtendedPubKey {
    pub network: Network,
    pub depth: u8,
    pub parent_fingerprint: Fingerprint,
    pub child_number: ChildNumber,
    pub public_key: PublicKey,
    pub chain_code: ChainCode,
}

Extended public key

Fields
network: Network

The network this key is to be used on

depth: u8

How many derivations this key is from the master (which is 0)

parent_fingerprint: Fingerprint

Fingerprint of the parent key

child_number: ChildNumber

Child number of the key used to derive from parent (0 for master)

public_key: PublicKey

Public key

chain_code: ChainCode

Chain code

Implementations

impl ExtendedPubKey

pub fn from_private<C: Signing>(
secp: &Secp256k1<C>,
sk: &ExtendedPrivKey
) -> ExtendedPubKey

👎Deprecated since 0.28.0: use ExtendedPubKey::from_priv

Derives a public key from a private key

pub fn from_priv<C: Signing>(
secp: &Secp256k1<C>,
sk: &ExtendedPrivKey
) -> ExtendedPubKey

Derives a public key from a private key

pub fn to_pub(self) -> PublicKey

Constructs ECDSA compressed public key matching internal public key representation.

pub fn to_x_only_pub(self) -> XOnlyPublicKey

Constructs BIP340 x-only public key for BIP-340 signatures and Taproot use matching the internal public key representation.

pub fn derive_pub<C: Verification, P: AsRef<[ChildNumber]>>(
&self,
secp: &Secp256k1<C>,
path: &P
) -> Result<ExtendedPubKey, Error>

Attempts to derive an extended public key from a path.

The path argument can be both of type DerivationPath or Vec<ChildNumber>.

pub fn ckd_pub_tweak(
&self,
i: ChildNumber
) -> Result<(SecretKey, ChainCode), Error>

Compute the scalar tweak added to this key to get a child key

pub fn ckd_pub<C: Verification>(
&self,
secp: &Secp256k1<C>,
i: ChildNumber
) -> Result<ExtendedPubKey, Error>

Public->Public child key derivation

pub fn decode(data: &[u8]) -> Result<ExtendedPubKey, Error>

Decoding extended public key from binary data according to BIP 32

pub fn encode(&self) -> [u8; 78]

Extended public key binary encoding according to BIP 32

pub fn identifier(&self) -> XpubIdentifier

Returns the HASH160 of the chaincode

pub fn fingerprint(&self) -> Fingerprint

Returns the first four bytes of the identifier