Module bitcoin::util::contracthash

👎 Deprecated

Pay-to-contract-hash support

See Appendix A of the Blockstream sidechains whitepaper at http://blockstream.com/sidechains.pdf for details of what this does.

Structs

TemplateDeprecated

A script template

Enums

ErrorDeprecated

A contract-hash error

Functions

compute_tweakDeprecated

Compute a tweak from some given data for the given public key

create_addressDeprecated

Takes a contract, template and key set and runs through all the steps

tweak_keyDeprecated

Tweak a single key using some arbitrary data

tweak_keysDeprecated

Tweak keys using some arbitrary data

tweak_secret_keyDeprecated

Tweak a secret key using some arbitrary data (calls compute_tweak internally)

untemplateDeprecated

Extract the keys and template from a completed script