Struct bitcoin::util::ecdsa::PrivateKey

pub struct PrivateKey {
    pub compressed: bool,
    pub network: Network,
    pub key: SecretKey,
}

A Bitcoin ECDSA private key

Fields
compressed: bool

Whether this private key should be serialized as compressed

network: Network

The network on which this key should be used

key: SecretKey

The actual ECDSA key

Implementations

impl PrivateKey

pub fn new(key: SecretKey, network: Network) -> PrivateKey

Constructs compressed ECDSA private key from the provided generic Secp256k1 private key and the specified network

pub fn new_uncompressed(key: SecretKey, network: Network) -> PrivateKey

Constructs uncompressed (legacy) ECDSA private key from the provided generic Secp256k1 private key and the specified network

pub fn public_key<C: Signing>(&self, secp: &Secp256k1<C>) -> PublicKey

Creates a public key from this private key

pub fn to_bytes(&self) -> Vec<u8>

Serialize the private key to bytes

pub fn from_slice(data: &[u8], network: Network) -> Result<PrivateKey, Error>

Deserialize a private key from a slice

pub fn fmt_wif(&self, fmt: &mut dyn Write) -> Result

Format the private key to WIF format.

pub fn to_wif(&self) -> String

Get WIF encoding of this private key.

pub fn from_wif(wif: &str) -> Result<PrivateKey, Error>

Parse WIF encoded private key.