Module bitcoin::util::key

Bitcoin keys.

This module provides keys used in Bitcoin that can be roundtrip (de)serialized.

Re-exports

pub use secp256k1;

Structs

KeyPair

Opaque data structure that holds a keypair consisting of a secret and a public key.

PrivateKey

A Bitcoin ECDSA private key

PublicKey

A Bitcoin ECDSA public key

Secp256k1

The secp256k1 engine, used to execute all signature operations.

SortKey

An opaque return type for PublicKey::to_sort_key

XOnlyPublicKey

An x-only public key, used for verification of Schnorr signatures and serialized according to BIP-340.

Enums

Error

A key-related error.