Struct bitcoin::util::key::XOnlyPublicKey

pub struct XOnlyPublicKey(_);

An x-only public key, used for verification of Schnorr signatures and serialized according to BIP-340.

Serde support

Implements de/serialization with the serde feature enabled. We treat the byte value as a tuple of 32 u8s for non-human-readable formats. This representation is optimal for for some formats (e.g. bincode) however other formats may be less optimal (e.g. cbor).

Examples

Basic usage:

use secp256k1::{rand, Secp256k1, KeyPair, XOnlyPublicKey};

let secp = Secp256k1::new();
let key_pair = KeyPair::new(&secp, &mut rand::thread_rng());
let xonly = XOnlyPublicKey::from_keypair(&key_pair);

Implementations

impl XOnlyPublicKey

pub fn as_ptr(&self) -> *const XOnlyPublicKey

Obtains a raw const pointer suitable for use with FFI functions.

pub fn as_mut_ptr(&mut self) -> *mut XOnlyPublicKey

Obtains a raw mutable pointer suitable for use with FFI functions.

pub fn from_keypair(keypair: &KeyPair) -> (XOnlyPublicKey, Parity)

Returns the XOnlyPublicKey (and it’s Parity) for keypair.

pub fn from_slice(data: &[u8]) -> Result<XOnlyPublicKey, Error>

Creates a Schnorr public key directly from a slice.

Errors

Returns Error::InvalidPublicKey if the length of the data slice is not 32 bytes or the slice does not represent a valid Secp256k1 point x coordinate.

pub fn serialize(&self) -> [u8; 32]

Serializes the key as a byte-encoded x coordinate value (32 bytes).

pub fn tweak_add_assign<V>(
&mut self,
secp: &Secp256k1<V>,
tweak: &Scalar
) -> Result<Parity, Error>where
V: Verification,

👎Deprecated since 0.23.0:
Use add_tweak instead

Tweaks an x-only PublicKey by adding the generator multiplied with the given tweak to it.

pub fn add_tweak<V>(
self,
secp: &Secp256k1<V>,
tweak: &Scalar
) -> Result<(XOnlyPublicKey, Parity), Error>where
V: Verification,

Tweaks an XOnlyPublicKey by adding the generator multiplied with the given tweak to it.

Returns

The newly tweaked key plus an opaque type representing the parity of the tweaked key, this should be provided to tweak_add_check which can be used to verify a tweak more efficiently than regenerating it and checking equality.

Errors

If the resulting key would be invalid.

Examples

use secp256k1::{Secp256k1, KeyPair, Scalar, XOnlyPublicKey};
use secp256k1::rand::{RngCore, thread_rng};

let secp = Secp256k1::new();
let tweak = Scalar::random();

let mut key_pair = KeyPair::new(&secp, &mut thread_rng());
let (xonly, _parity) = key_pair.x_only_public_key();
let tweaked = xonly.add_tweak(&secp, &tweak).expect("Improbable to fail with a randomly generated tweak");

pub fn tweak_add_check<V>(
&self,
secp: &Secp256k1<V>,
tweaked_key: &XOnlyPublicKey,
tweaked_parity: Parity,
tweak: Scalar
) -> boolwhere
V: Verification,

Verifies that a tweak produced by XOnlyPublicKey::tweak_add_assign was computed correctly.

Should be called on the original untweaked key. Takes the tweaked key and output parity from XOnlyPublicKey::tweak_add_assign as input.

Currently this is not much more efficient than just recomputing the tweak and checking equality. However, in future this API will support batch verification, which is significantly faster, so it is wise to design protocols with this in mind.

Returns

True if tweak and check is successful, false otherwise.

Examples

use secp256k1::{Secp256k1, KeyPair, Scalar};
use secp256k1::rand::{thread_rng, RngCore};

let secp = Secp256k1::new();
let tweak = Scalar::random();

let mut key_pair = KeyPair::new(&secp, &mut thread_rng());
let (mut public_key, _) = key_pair.x_only_public_key();
let original = public_key;
let parity = public_key.tweak_add_assign(&secp, &tweak).expect("Improbable to fail with a randomly generated tweak");
assert!(original.tweak_add_check(&secp, &public_key, parity, tweak));

pub fn public_key(&self, parity: Parity) -> PublicKey

Returns the PublicKey for this XOnlyPublicKey.

This is equivalent to using [PublicKey::from_xonly_and_parity(self, parity)].