Module bitcoin::util::misc

Miscellaneous functions.

This module provides various utility functions including secp256k1 signature recovery when library is used with the secp-recovery feature.

Structs

MessageSignaturesecp-recovery

A signature on a Bitcoin Signed Message.

Enums

MessageSignatureErrorsecp-recovery

An error used for dealing with Bitcoin Signed Messages.

Constants

BITCOIN_SIGNED_MSG_PREFIX

The prefix for signed messages using Bitcoin’s message signing protocol.

Functions

script_find_and_remove

Search for needle in the vector haystack and remove every instance of it, returning the number of instances removed. Loops through the vector opcode by opcode, skipping pushed data.

signed_msg_hash

Hash message for signature using Bitcoin’s message signing format.