Struct bitcoin::util::psbt::Input

pub struct Input {

    pub non_witness_utxo: Option<Transaction>,
    pub witness_utxo: Option<TxOut>,
    pub partial_sigs: BTreeMap<PublicKey, Vec<u8>>,
    pub sighash_type: Option<SigHashType>,
    pub redeem_script: Option<Script>,
    pub witness_script: Option<Script>,
    pub bip32_derivation: BTreeMap<PublicKey, KeySource>,
    pub final_script_sig: Option<Script>,
    pub final_script_witness: Option<Vec<Vec<u8>>>,
    pub ripemd160_preimages: BTreeMap<Hash, Vec<u8>>,
    pub sha256_preimages: BTreeMap<Hash, Vec<u8>>,
    pub hash160_preimages: BTreeMap<Hash, Vec<u8>>,
    pub hash256_preimages: BTreeMap<Hash, Vec<u8>>,
    pub proprietary: BTreeMap<ProprietaryKey, Vec<u8>>,
    pub unknown: BTreeMap<Key, Vec<u8>>,

}

A key-value map for an input of the corresponding index in the unsigned transaction.

Fields
non_witness_utxo: Option<Transaction>

The non-witness transaction this input spends from. Should only be std::option::Option::Some for inputs which spend non-segwit outputs or if it is unknown whether an input spends a segwit output.

witness_utxo: Option<TxOut>

The transaction output this input spends from. Should only be std::option::Option::Some for inputs which spend segwit outputs, including P2SH embedded ones.

partial_sigs: BTreeMap<PublicKey, Vec<u8>>

A map from public keys to their corresponding signature as would be pushed to the stack from a scriptSig or witness.

sighash_type: Option<SigHashType>

The sighash type to be used for this input. Signatures for this input must use the sighash type.

redeem_script: Option<Script>

The redeem script for this input.

witness_script: Option<Script>

The witness script for this input.

bip32_derivation: BTreeMap<PublicKey, KeySource>

A map from public keys needed to sign this input to their corresponding master key fingerprints and derivation paths.

final_script_sig: Option<Script>

The finalized, fully-constructed scriptSig with signatures and any other scripts necessary for this input to pass validation.

final_script_witness: Option<Vec<Vec<u8>>>

The finalized, fully-constructed scriptWitness with signatures and any other scripts necessary for this input to pass validation.

ripemd160_preimages: BTreeMap<Hash, Vec<u8>>

TODO: Proof of reserves commitment RIPEMD160 hash to preimage map

sha256_preimages: BTreeMap<Hash, Vec<u8>>

SHA256 hash to preimage map

hash160_preimages: BTreeMap<Hash, Vec<u8>>

HSAH160 hash to preimage map

hash256_preimages: BTreeMap<Hash, Vec<u8>>

HAS256 hash to preimage map

proprietary: BTreeMap<ProprietaryKey, Vec<u8>>

Proprietary key-value pairs for this input.

unknown: BTreeMap<Key, Vec<u8>>

Unknown key-value pairs for this input.