Struct bitcoin::util::psbt::PartiallySignedTransaction

pub struct PartiallySignedTransaction {
    pub unsigned_tx: Transaction,
    pub version: u32,
    pub xpub: BTreeMap<ExtendedPubKey, KeySource>,
    pub proprietary: BTreeMap<ProprietaryKey, Vec<u8>>,
    pub unknown: BTreeMap<Key, Vec<u8>>,
    pub inputs: Vec<Input>,
    pub outputs: Vec<Output>,
}

A Partially Signed Transaction.

Fields
unsigned_tx: Transaction

The unsigned transaction, scriptSigs and witnesses for each input must be empty.

version: u32

The version number of this PSBT. If omitted, the version number is 0.

xpub: BTreeMap<ExtendedPubKey, KeySource>

A global map from extended public keys to the used key fingerprint and derivation path as defined by BIP 32.

proprietary: BTreeMap<ProprietaryKey, Vec<u8>>

Global proprietary key-value pairs.

unknown: BTreeMap<Key, Vec<u8>>

Unknown global key-value pairs.

inputs: Vec<Input>

The corresponding key-value map for each input in the unsigned transaction.

outputs: Vec<Output>

The corresponding key-value map for each output in the unsigned transaction.

Implementations

impl PartiallySignedTransaction

pub fn iter_funding_utxos(&self) -> impl Iterator<Item = Result<&TxOut, Error>>

Returns an iterator for the funding UTXOs of the psbt

For each PSBT input that contains UTXO information Ok is returned containing that information. The order of returned items is same as the order of inputs.

Errors

The function returns error when UTXO information is not present or is invalid.

Panics

The function panics if the length of transaction inputs is not equal to the length of PSBT inputs.

pub fn from_unsigned_tx(tx: Transaction) -> Result<Self, Error>

Creates a PSBT from an unsigned transaction.

Errors

If transactions is not unsigned.

pub fn extract_tx(self) -> Transaction

Extracts the Transaction from a PSBT by filling in the available signature information.

pub fn combine(&mut self, other: Self) -> Result<(), Error>

Combines this PartiallySignedTransaction with other PSBT as described by BIP 174.

In accordance with BIP 174 this function is commutative i.e., A.combine(B) == B.combine(A)