Struct bitcoin::util::schnorr::PublicKey

pub struct PublicKey(_);

A Schnorr public key, used for verification of Schnorr signatures

Implementations

impl PublicKey

pub fn as_ptr(&self) -> *const XOnlyPublicKey

Obtains a raw const pointer suitable for use with FFI functions

pub fn as_mut_ptr(&mut self) -> *mut XOnlyPublicKey

Obtains a raw mutable pointer suitable for use with FFI functions

pub fn from_keypair<C>(secp: &Secp256k1<C>, keypair: &KeyPair) -> PublicKey where
C: Signing,

Creates a new Schnorr public key from a Schnorr key pair

pub fn from_slice(data: &[u8]) -> Result<PublicKey, Error>

Creates a Schnorr public key directly from a slice

pub fn serialize(&self) -> [u8; 32]

Serialize the key as a byte-encoded pair of values. In compressed form the y-coordinate is represented by only a single bit, as x determines it up to one bit.

pub fn tweak_add_assign<V>(
&mut self,
secp: &Secp256k1<V>,
tweak: &[u8]
) -> Result<bool, Error> where
V: Verification,

Tweak an x-only PublicKey by adding the generator multiplied with the given tweak to it.

Returns a boolean representing the parity of the tweaked key, which can be provided to tweak_add_check which can be used to verify a tweak more efficiently than regenerating it and checking equality. Will return an error if the resulting key would be invalid or if the tweak was not a 32-byte length slice.

pub fn tweak_add_check<V>(
&self,
secp: &Secp256k1<V>,
tweaked_key: &PublicKey,
tweaked_parity: bool,
tweak: [u8; 32]
) -> bool where
V: Verification,

Verify that a tweak produced by tweak_add_assign was computed correctly

Should be called on the original untweaked key. Takes the tweaked key and output parity from tweak_add_assign as input.

Currently this is not much more efficient than just recomputing the tweak and checking equality. However, in future this API will support batch verification, which is significantly faster, so it is wise to design protocols with this in mind.