Module bitcoin::util::sighash

Generalized, efficient, signature hash implementation.

Implementation of the algorithm to compute the message to be signed according to Bip341, Bip143 and legacy (before Bip143).

Re-exports

pub use blockdata::transaction::EcdsaSighashType;

pub use blockdata::transaction::SighashTypeParseError;

Structs

Annex

The Annex struct is a slice wrapper enforcing first byte to be 0x50.

ScriptPath

Information related to the script path spending

SighashCache

Efficiently calculates signature hash message for legacy, segwit and taproot inputs.

Enums

Error

Possible errors in computing the signature message

Prevouts

Contains outputs of previous transactions. In the case SchnorrSighashType variant is ANYONECANPAY, Prevouts::One may be provided

SchnorrSighashType

Hashtype of an input’s signature, encoded in the last byte of the signature Fixed values so they can be casted as integer types for encoding