Struct bitcoin::util::taproot::TaprootBuilder

pub struct TaprootBuilder { /* private fields */ }

Builder for building taproot iteratively. Users can specify tap leaf or omitted/hidden branches in a depth-first search (DFS) walk order to construct this tree.

See Wikipedia for more details on DFS.

Implementations

impl TaprootBuilder

pub fn new() -> Self

Creates a new instance of TaprootBuilder.

pub fn with_huffman_tree<I>(
script_weights: I
) -> Result<Self, TaprootBuilderError>where
I: IntoIterator<Item = (u32, Script)>,

Creates a new TaprootSpendInfo from a list of scripts (with default script version) and weights of satisfaction for that script.

The weights represent the probability of each branch being taken. If probabilities/weights for each condition are known, constructing the tree as a Huffman Tree is the optimal way to minimize average case satisfaction cost. This function takes as input an iterator of tuple(u32, &Script) where u32 represents the satisfaction weights of the branch. For example, [(3, S1), (2, S2), (5, S3)] would construct a TapTree that has optimal satisfaction weight when probability for S1 is 30%, S2 is 20% and S3 is 50%.

Errors:

When the optimal Huffman Tree has a depth more than 128.

If the provided list of script weights is empty.

Edge Cases:

If the script weight calculations overflow, a sub-optimal tree may be generated. This should not happen unless you are dealing with billions of branches with weights close to 2^32.

pub fn add_leaf_with_ver(
self,
depth: u8,
script: Script,
ver: LeafVersion
) -> Result<Self, TaprootBuilderError>

Adds a leaf script at depth to the builder with script version ver. Errors if the leaves are not provided in DFS walk order. The depth of the root node is 0.

pub fn add_leaf(
self,
depth: u8,
script: Script
) -> Result<Self, TaprootBuilderError>

Adds a leaf script at depth to the builder with default script version. Errors if the leaves are not provided in DFS walk order. The depth of the root node is 0.

See TaprootBuilder::add_leaf_with_ver for adding a leaf with specific version.

pub fn add_hidden_node(
self,
depth: u8,
hash: Hash
) -> Result<Self, TaprootBuilderError>

Adds a hidden/omitted node at depth to the builder. Errors if the leaves are not provided in DFS walk order. The depth of the root node is 0.

pub fn is_finalizable(&self) -> bool

Checks if the builder has finalized building a tree.

pub fn has_hidden_nodes(&self) -> bool

Checks if the builder has hidden nodes.

pub fn finalize<C: Verification>(
self,
secp: &Secp256k1<C>,
internal_key: UntweakedPublicKey
) -> Result<TaprootSpendInfo, TaprootBuilder>

Creates a TaprootSpendInfo with the given internal key.

Returns the unmodified builder as Err if the builder is not finalizable. See also TaprootBuilder::is_finalizable