Struct bitcoin::util::uint::Uint256

pub struct Uint256(pub [u64; 4]);

Little-endian large integer type

Tuple Fields
0: [u64; 4]
Implementations

impl Uint256

pub fn as_ptr(&self) -> *const u64

Converts the object to a raw pointer

pub fn as_mut_ptr(&mut self) -> *mut u64

Converts the object to a mutable raw pointer

pub fn len(&self) -> usize

Returns the length of the object as an array

pub fn is_empty(&self) -> bool

Returns whether the object, as an array, is empty. Always false.

pub fn as_bytes(&self) -> &[u64; 4]

Returns the underlying bytes.

pub fn to_bytes(&self) -> [u64; 4]

Returns the underlying bytes.

pub fn into_bytes(self) -> [u64; 4]

Returns the underlying bytes.

impl Uint256

pub fn low_u32(&self) -> u32

Conversion to u32

pub fn low_u64(&self) -> u64

Conversion to u64

pub fn bits(&self) -> usize

Return the least number of bits needed to represent the number

pub fn mul_u32(self, other: u32) -> Uint256

Multiplication by u32

pub fn from_u64(init: u64) -> Option<Uint256>

Create an object from a given unsigned 64-bit integer

pub fn from_i64(init: i64) -> Option<Uint256>

Create an object from a given signed 64-bit integer

pub fn from_be_bytes(bytes: [u8; 32]) -> Uint256

Creates big integer value from a byte array using big-endian encoding

pub fn from_be_slice(bytes: &[u8]) -> Result<Uint256, ParseLengthError>

Creates big integer value from a byte slice using big-endian encoding

pub fn to_be_bytes(&self) -> [u8; 32]

Convert a big integer into a byte array using big-endian encoding

impl Uint256

pub fn increment(&mut self)

Increment by 1

pub fn low_128(&self) -> Uint128

Decay to a uint128