Crate bitcoin

Bitcoin Library

This crate provides low-level parsing and verification functionality for Bitcoin specific data structures.

Unless otherwise stated, the primary source of truth for code contained herein is the Bitcoin Core repository, though implementation details may vary.

Overview

This crate provides functions for:

(De)serialization of block headers, transactions and merkle proofs.

Script (address) construction and parsing.

Merkle proof construction and verification.

Elliptic curve multiplication over Secp256k1.

Modules

formatter

global

merkle

parser

types

utils

Structs

PublicKey

Compressed ECDSA (secp256k1 curve) Public Key

Script

Bitcoin script

Enums

Address

A Bitcoin address is a serialized identifier that represents the destination for a payment. Address prefixes are used to indicate the network as well as the format. Since the Parachain follows SPV assumptions we do not need to know which network a payment is included in.

Error