Crate bitcoin

Re-exports

pub use light::BitcoinLight;

pub use bitcoincore_rpc::bitcoin as bitcoin_primitives;

pub use bitcoincore_rpc::bitcoin::hashes;

pub use bitcoincore_rpc::bitcoin::secp256k1;

pub use bitcoincore_rpc::json;

pub use bitcoincore_rpc::jsonrpc;

Modules

address

Bitcoin addresses.

cli

consensus

Bitcoin consensus.

key

Bitcoin keys.

light

opcodes

Enables wildcard imports to bring into scope all opcodes and nothing else.

psbt

Partially Signed Bitcoin Transactions.

sha256

SHA256 implementation.

util

Utility functions.

Structs

Address

A Bitcoin address.

Amount

Amount

BitcoinCore

BitcoinCoreBuilder

Block

Bitcoin block.

BlockHash

A bitcoin block hash.

BlockHeader

Bitcoin block header.

Builder

An Object which can be used to construct a script piece by piece.

Client

Client implements a JSON-RPC client for the Bitcoin Core daemon or compatible APIs.

CreateRawTransactionInput

EcdsaSig

An ECDSA signature with the corresponding hash type.

ElectrsClient

FundRawTransactionOptions

GetBlockResult

GetBlockchainInfoResult

Models the result of “getblockchaininfo”

GetRawTransactionResult

GetTransactionResult

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

LockedTransaction

NonStandardSighashType

This type is consensus valid but an input including it would prevent the transaction from being relayed on today’s Bitcoin network.

OutPoint

A reference to a transaction output.

PartialMerkleTree

Data structure that represents a partial merkle tree.

PrivateKey

A Bitcoin ECDSA private key

PubkeyHash

A hash of a public key.

PublicKey

A Bitcoin ECDSA public key

RawTransactionProof

RpcError

A JSONRPC error object

SatPerVbyte

Script

Bitcoin script slice.

ScriptHash

A hash of Bitcoin Script bytecode.

SecretKey

Secret 256-bit key used as x in an ECDSA signature.

SignedAmount

SignedAmount

Transaction

Bitcoin transaction.

TransactionMetadata

TxIn

Bitcoin transaction input.

TxMerkleNode

A hash of the Merkle tree branch or root for transactions

TxOut

Bitcoin transaction output.

Txid

A bitcoin transaction hash/transaction ID.

VarInt

A variable-length unsigned integer.

WPubkeyHash

SegWit version of a public key hash.

WScriptHash

SegWit version of a Bitcoin Script bytecode hash.

WalletTxInfo

Enums

AddressType

Used to represent an address type.

Auth

The different authentication methods for the client.

BitcoinError

The error type for errors produced in this library.

BitcoinLightError

BitcoinRpcError

ConversionError

ElectrsError

Error

GetTransactionResultDetailCategory

Enum to represent the category of a transaction.

JsonRpcError

A library error

LockTime

An absolute lock time value, representing either a block height or a UNIX timestamp (seconds since epoch).

Network

The cryptocurrency network to act on.

Payload

The method used to produce an address.

Constants

BITCOIN_CORE_VERSION_23

the bitcoin core version. See https://github.com/bitcoin/bitcoin/blob/833add0f48b0fad84d7b8cf9373a349e7aef20b4/src/rpc/net.cpp#L627 and https://github.com/bitcoin/bitcoin/blob/833add0f48b0fad84d7b8cf9373a349e7aef20b4/src/clientversion.h#L33-L37

BLOCK_INTERVAL

Average time to mine a Bitcoin block.

DEFAULT_MAX_TX_COUNT

PUBLIC_KEY_SIZE

The size (in bytes) of a serialized public key.

TRANSPORT_TIMEOUT

timeout on the jsonrpc transport. jsonrpc default is 15 seconds.

Traits

BitcoinCoreApi

FromHex

Trait for objects that can be deserialized from hex strings.

Hash

Trait which applies to hashes of all types.

RpcApi

TransactionExt

Extension trait for transaction, adding methods to help to match the Transaction to Replace/Redeem requests

Functions

deserialize

Deserializes an object from a vector, will error if said deserialization doesn’t consume the entire vector.

reverse_stream_transactions

Stream over transactions, starting with this in the mempool and continuing with transactions from previous in-chain block. The stream ends after the block at stop_height has been returned.

serialize

Encodes an object into a vector.

stream_blocks

Stream blocks continuously from_height awaiting the production of new blocks as reported by Bitcoin core. The stream never ends.

stream_in_chain_transactions

Stream all transactions in blocks produced by Bitcoin Core.