Crate bitcoin

Rust Bitcoin Library

This is a library for which supports the Bitcoin network protocol and associated primitives. It is designed for Rust programs built to work with the Bitcoin network.

It is also written entirely in Rust to illustrate the benefits of strong type safety, including ownership and lifetime, for financial and/or cryptographic software.

Re-exports

pub extern crate bech32;

pub extern crate bitcoin_hashes as hashes;

pub extern crate secp256k1;

pub use hash_types::*;

pub use blockdata::block::Block;

pub use blockdata::block::BlockHeader;

pub use blockdata::script::Script;

pub use blockdata::transaction::Transaction;

pub use blockdata::transaction::TxIn;

pub use blockdata::transaction::TxOut;

pub use blockdata::transaction::OutPoint;

pub use blockdata::transaction::SigHashType;

pub use consensus::encode::VarInt;

pub use network::constants::Network;

pub use util::Error;

pub use util::address::Address;

pub use util::address::AddressType;

pub use util::amount::Amount;

pub use util::amount::Denomination;

pub use util::amount::SignedAmount;

pub use util::merkleblock::MerkleBlock;

pub use util::ecdsa;

pub use util::schnorr;

pub use util::ecdsa::PrivateKey;

pub use util::ecdsa::PublicKey;

Modules

blockdata
Blockdata

consensus
Consensus

hash_types
File defines types for hashes used throughout the library. These types are needed in order to avoid mixing data of the same hash format (like SHA256d) but of different meaning (transaction id, block hash etc).

network
Network Support

policy
Policy

util
Utility functions