bitcoin::address

Enum Error

#[non_exhaustive]
pub enum Error {

    Base58(Error),
    Bech32(Error),
    EmptyBech32Payload,
    InvalidBech32Variant {
        expected: Variant,
        found: Variant,
    },
    InvalidWitnessVersion(u8),
    UnparsableWitnessVersion(ParseIntError),
    MalformedWitnessVersion,
    InvalidWitnessProgramLength(usize),
    InvalidSegwitV0ProgramLength(usize),
    UncompressedPubkey,
    ExcessiveScriptSize,
    UnrecognizedScript,
    UnknownAddressType(String),
    NetworkValidation {
        required: Network,
        found: Network,
        address: Address<NetworkUnchecked>,
    },

}

Address error.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Base58(Error)

Base58 encoding error.

Bech32(Error)

Bech32 encoding error.

EmptyBech32Payload

The bech32 payload was empty.

InvalidBech32Variant

The wrong checksum algorithm was used. See BIP-0350.

Fields

expected: Variant

Bech32 variant that is required by the used Witness version.

found: Variant

The actual Bech32 variant encoded in the address representation.

InvalidWitnessVersion(u8)

Script version must be 0 to 16 inclusive.

UnparsableWitnessVersion(ParseIntError)

Unable to parse witness version from string.

MalformedWitnessVersion

Bitcoin script opcode does not match any known witness version, the script is malformed.

InvalidWitnessProgramLength(usize)

The witness program must be between 2 and 40 bytes in length.

InvalidSegwitV0ProgramLength(usize)

A v0 witness program must be either of length 20 or 32.

UncompressedPubkey

An uncompressed pubkey was used where it is not allowed.

ExcessiveScriptSize

Address size more than 520 bytes is not allowed.

UnrecognizedScript

Script is not a p2pkh, p2sh or witness program.

UnknownAddressType(String)

Address type is either invalid or not supported in rust-bitcoin.

NetworkValidation

Address’s network differs from required one.

Fields

required: Network

Network that was required.

found: Network

Network on which the address was found to be valid.

address: Address<NetworkUnchecked>

The address itself