Module bitcoin::bip32

BIP32 implementation.

Implementation of BIP32 hierarchical deterministic wallets, as defined at https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki.

Structs

ChainCode

A chain code

DerivationPath

A BIP-32 derivation path.

DerivationPathIterator

An iterator over children of a DerivationPath.

ExtendedPrivKey

Extended private key

ExtendedPubKey

Extended public key

Fingerprint

A fingerprint

Enums

ChildNumber

A child number for a derived key

Error

A BIP32 error

Traits

IntoDerivationPath

Trait that allows possibly failable conversion from a type into a derivation path

Type Definitions

KeySource

Full information on the used extended public key: fingerprint of the master extended public key and a derivation path from it.