Module bitcoin::blockdata

Bitcoin block data.

This module defines structures and functions for storing the blocks and transactions which make up the Bitcoin system.

Modules

block

Bitcoin blocks.

constants

Blockdata constants.

locktime

Provides type LockTime that implements the logic around nLockTime/OP_CHECKLOCKTIMEVERIFY.

opcodes

Bitcoin script opcodes.

script

Bitcoin scripts.

transaction

Bitcoin transactions.

witness

Witness