Module bitcoin::blockdata

Bitcoin block data.

This module defines structures and functions for storing the blocks and transactions which make up the Bitcoin system.

Re-exports

pub use fee_rate::FeeRate;

pub use weight::Weight;

Modules

block

Bitcoin blocks.

constants

Blockdata constants.

fee_rate

Implements FeeRate and assoctiated features.

locktime

Provides absolute and relative locktimes.

opcodes

Bitcoin script opcodes.

script

Bitcoin scripts.

transaction

Bitcoin transactions.

weight

Implements Weight and associated features.

witness

Witness