Module bitcoin::blockdata

Blockdata

This module defines structures and functions for storing the blocks and transactions which make up the Bitcoin system.

Modules

block

Bitcoin Block

constants

Blockdata constants

opcodes

Opcodes

script

Script

transaction

Bitcoin Transaction