Struct bitcoin::blockdata::block::Block

pub struct Block {
    pub header: BlockHeader,
    pub txdata: Vec<Transaction>,
}

A Bitcoin block, which is a collection of transactions with an attached proof of work.

Fields
header: BlockHeader

The block header

txdata: Vec<Transaction>

List of transactions contained in the block

Implementations

impl Block

pub fn block_hash(&self) -> BlockHash

Return the block hash.

pub fn check_merkle_root(&self) -> bool

check if merkle root of header matches merkle root of the transaction list

pub fn check_witness_commitment(&self) -> bool

check if witness commitment in coinbase is matching the transaction list

pub fn merkle_root(&self) -> TxMerkleNode

Calculate the transaction merkle root.

pub fn compute_witness_commitment(
witness_root: &WitnessMerkleNode,
witness_reserved_value: &[u8]
) -> WitnessCommitment

compute witness commitment for the transaction list

pub fn witness_root(&self) -> WitnessMerkleNode

Merkle root of transactions hashed for witness

pub fn get_size(&self) -> usize

Get the size of the block

pub fn get_weight(&self) -> usize

Get the weight of the block

pub fn coinbase(&self) -> Option<&Transaction>

Get the coinbase transaction, if one is present.

pub fn bip34_block_height(&self) -> Result<u64, Bip34Error>

Get the block height as encoded into the coinbase according to BIP34. Returns None if not present.