Struct bitcoin::blockdata::script::Script

pub struct Script(_);

Bitcoin script.

A list of instructions in a simple, Forth-like, stack-based programming language that Bitcoin uses.

See Bitcoin Wiki: Script for more information.

Bitcoin Core References

CScript definition

Implementations

impl Script

pub fn new() -> Script

Creates a new empty script.

pub fn new_p2pk(pubkey: &PublicKey) -> Script

Generates P2PK-type of scriptPubkey.

pub fn new_p2pkh(pubkey_hash: &PubkeyHash) -> Script

Generates P2PKH-type of scriptPubkey.

pub fn new_p2sh(script_hash: &ScriptHash) -> Script

Generates P2SH-type of scriptPubkey with a given hash of the redeem script.

pub fn new_v0_wpkh(pubkey_hash: &WPubkeyHash) -> Script

👎 Deprecated since 0.28.0:
use Script::new_v0_p2wpkh method instead

Generates P2WPKH-type of scriptPubkey.

pub fn new_v0_p2wpkh(pubkey_hash: &WPubkeyHash) -> Script

Generates P2WPKH-type of scriptPubkey.

pub fn new_v0_wsh(script_hash: &WScriptHash) -> Script

👎 Deprecated since 0.28.0:
use Script::new_v0_p2wsh method instead

Generates P2WSH-type of scriptPubkey with a given hash of the redeem script.

pub fn new_v0_p2wsh(script_hash: &WScriptHash) -> Script

Generates P2WSH-type of scriptPubkey with a given hash of the redeem script.

pub fn new_v1_p2tr<C: Verification>(
secp: &Secp256k1<C>,
internal_key: UntweakedPublicKey,
merkle_root: Option<TapBranchHash>
) -> Script

Generates P2TR for script spending path using an internal public key and some optional script tree merkle root.

pub fn new_v1_p2tr_tweaked(output_key: TweakedPublicKey) -> Script

Generates P2TR for key spending path for a known TweakedPublicKey.

pub fn new_witness_program(version: WitnessVersion, program: &[u8]) -> Script

Generates P2WSH-type of scriptPubkey with a given hash of the redeem script.

pub fn new_op_return(data: &[u8]) -> Script

Generates OP_RETURN-type of scriptPubkey for the given data.

pub fn script_hash(&self) -> ScriptHash

Returns 160-bit hash of the script.

pub fn wscript_hash(&self) -> WScriptHash

Returns 256-bit hash of the script for P2WSH outputs.

pub fn len(&self) -> usize

Returns the length in bytes of the script.

pub fn is_empty(&self) -> bool

Returns whether the script is the empty script.

pub fn as_bytes(&self) -> &[u8]

Returns the script data as a byte slice.

pub fn to_bytes(&self) -> Vec<u8>

Returns a copy of the script data.

pub fn into_bytes(self) -> Vec<u8>

Converts the script into a byte vector.

pub fn to_p2sh(&self) -> Script

Computes the P2SH output corresponding to this redeem script.

pub fn p2wpkh_script_code(&self) -> Option<Script>

Returns the script code used for spending a P2WPKH output if this script is a script pubkey for a P2WPKH output. The scriptCode is described in BIP143.

pub fn to_v0_p2wsh(&self) -> Script

Computes the P2WSH output corresponding to this witnessScript (aka the “witness redeem script”).

pub fn to_v1_p2tr<C: Verification>(
&self,
secp: &Secp256k1<C>,
internal_key: UntweakedPublicKey
) -> Script

Computes P2TR output with a given internal key and a single script spending path equal to the current script, assuming that the script is a Tapscript.

pub fn witness_version(&self) -> Option<WitnessVersion>

Returns witness version of the script, if any, assuming the script is a scriptPubkey.

pub fn is_p2sh(&self) -> bool

Checks whether a script pubkey is a P2SH output.

pub fn is_p2pkh(&self) -> bool

Checks whether a script pubkey is a P2PKH output.

pub fn is_p2pk(&self) -> bool

Checks whether a script pubkey is a P2PK output.

pub fn is_witness_program(&self) -> bool

Checks whether a script pubkey is a Segregated Witness (segwit) program.

pub fn is_v0_p2wsh(&self) -> bool

Checks whether a script pubkey is a P2WSH output.

pub fn is_v0_p2wpkh(&self) -> bool

Checks whether a script pubkey is a P2WPKH output.

pub fn is_v1_p2tr(&self) -> bool

Checks whether a script pubkey is a P2TR output.

pub fn is_op_return(&self) -> bool

Check if this is an OP_RETURN output.

pub fn is_provably_unspendable(&self) -> bool

Checks whether a script can be proven to have no satisfying input.

pub fn dust_value(&self) -> Amount

Returns the minimum value an output with this script should have in order to be broadcastable on today’s Bitcoin network.

pub fn instructions(&self) -> Instructions<'_>

Iterates over the script in the form of Instructions, which are an enum covering opcodes, datapushes and errors.

At most one error will be returned and then the iterator will end. To instead iterate over the script as sequence of bytes, treat it as a slice using script[..] or convert it to a vector using into_bytes().

To force minimal pushes, use Self::instructions_minimal.

pub fn instructions_minimal(&self) -> Instructions<'_>

Iterates over the script in the form of Instructions while enforcing minimal pushes.

pub fn bytes_to_asm_fmt(script: &[u8], f: &mut dyn Write) -> Result

Writes the assembly decoding of the script bytes to the formatter.

pub fn fmt_asm(&self, f: &mut dyn Write) -> Result

Writes the assembly decoding of the script to the formatter.

pub fn bytes_to_asm(script: &[u8]) -> String

Creates an assembly decoding of the script in the given byte slice.

pub fn asm(&self) -> String

Returns the assembly decoding of the script.