Module bitcoin::blockdata::transaction

Bitcoin transactions.

A transaction describes a transfer of money. It consumes previously-unspent transaction outputs and produces new ones, satisfying the condition to spend the old outputs (typically a digital signature with a specific key must be provided) and defining the condition to spend the new ones. The use of digital signatures ensures that coins cannot be spent by unauthorized parties.

This module provides the structures and functions needed to support transactions.

Structs

NonStandardSighashType

This type is consensus valid but an input including it would prevent the transaction from being relayed on today’s Bitcoin network.

OutPoint

A reference to a transaction output.

Sequence

Bitcoin transaction input sequence number.

SighashTypeParseError

Error returned for failure during parsing one of the sighash types.

Transaction

Bitcoin transaction.

TxIn

Bitcoin transaction input.

TxOut

Bitcoin transaction output.

Enums

EcdsaSighashType

Hashtype of an input’s signature, encoded in the last byte of the signature.

EncodeSigningDataResult

Result of Transaction::encode_signing_data_to.

ParseOutPointError

An error in parsing an OutPoint.

RelativeLockTimeError

An error in creating relative lock-times.

Type Definitions

SigHashTypeDeprecated

Legacy Hashtype of an input’s signature