Module bitcoin::blockdata::transaction

Bitcoin Transaction

A transaction describes a transfer of money. It consumes previously-unspent transaction outputs and produces new ones, satisfying the condition to spend the old outputs (typically a digital signature with a specific key must be provided) and defining the condition to spend the new ones. The use of digital signatures ensures that coins cannot be spent by unauthorized parties.

This module provides the structures and functions needed to support transactions.

Structs

NonStandardSigHashType

This type is consensus valid but an input including it would prevent the transaction from being relayed on today’s Bitcoin network.

OutPoint

A reference to a transaction output

Transaction

A Bitcoin transaction, which describes an authenticated movement of coins.

TxIn

A transaction input, which defines old coins to be consumed

TxOut

A transaction output, which defines new coins to be created from old ones.

Enums

ParseOutPointError

An error in parsing an OutPoint.

SigHashType

Hashtype of an input’s signature, encoded in the last byte of the signature Fixed values so they can be casted as integer types for encoding