bitcoin::blockdata::witness

Struct Witness

pub struct Witness { /* private fields */ }

The Witness is the data used to unlock bitcoin since the segwit upgrade.

Can be logically seen as an array of bytestrings, i.e. Vec<Vec<u8>>, and it is serialized on the wire in that format. You can convert between this type and Vec<Vec<u8>> by using Witness::from_slice and Witness::to_vec.

For serialization and deserialization performance it is stored internally as a single Vec, saving some allocations.

Implementations

impl Witness

pub fn new() -> Self

Creates a new empty Witness.

pub fn from_vec(vec: Vec<Vec<u8>>) -> Self

👎Deprecated since 0.30.0: use Witness::from_slice() instead

Creates Witness object from an array of byte-arrays

pub fn from_slice<T: AsRef<[u8]>>(slice: &[T]) -> Self

Creates a Witness object from a slice of bytes slices where each slice is a witness item.

pub fn to_vec(&self) -> Vec<Vec<u8>>

Convenience method to create an array of byte-arrays from this witness.

pub fn is_empty(&self) -> bool

Returns true if the witness contains no element.

pub fn iter(&self) -> Iter<'_>

Returns a struct implementing Iterator.

pub fn len(&self) -> usize

Returns the number of elements this witness holds.

pub fn serialized_len(&self) -> usize

Returns the bytes required when this Witness is consensus encoded.

pub fn clear(&mut self)

Clear the witness.

pub fn push<T: AsRef<[u8]>>(&mut self, new_element: T)

Push a new element on the witness, requires an allocation.

pub fn push_bitcoin_signature( &mut self, signature: &SerializedSignature, hash_type: EcdsaSighashType, )

Pushes a DER-encoded ECDSA signature with a signature hash type as a new element on the witness, requires an allocation.

pub fn last(&self) -> Option<&[u8]>

Returns the last element in the witness, if any.

pub fn second_to_last(&self) -> Option<&[u8]>

Returns the second-to-last element in the witness, if any.

pub fn nth(&self, index: usize) -> Option<&[u8]>

Return the nth element in the witness, if any

pub fn tapscript(&self) -> Option<&Script>

Get Tapscript following BIP341 rules regarding accounting for an annex.

This does not guarantee that this represents a P2TR Witness. It merely gets the second to last or third to last element depending on the first byte of the last element being equal to 0x50. See Script::is_v1_p2tr to check whether this is actually a Taproot witness.