bitcoin

Module hash_types

Bitcoin hash types.

This module defines types for hashes used throughout the library. These types are needed in order to avoid mixing data of the same hash format (e.g. SHA256d) but of different meaning (such as transaction id, block hash).

Structs

BlockHash

A bitcoin block hash.

FilterHash

Filter hash, as defined in BIP-157

FilterHeader

Filter header, as defined in BIP-157

PubkeyHash

A hash of a public key.

ScriptHash

A hash of Bitcoin Script bytecode.

TxMerkleNode

A hash of the Merkle tree branch or root for transactions

Txid

A bitcoin transaction hash/transaction ID.

WPubkeyHash

SegWit version of a public key hash.

WScriptHash

SegWit version of a Bitcoin Script bytecode hash.

WitnessCommitment

A hash corresponding to the witness structure commitment in the coinbase transaction

WitnessMerkleNode

A hash corresponding to the Merkle tree root for witness data

Wtxid

A bitcoin witness transaction ID.

XpubIdentifier

XpubIdentifier as defined in BIP-32.