Module bitcoin::network::address

Bitcoin network addresses.

This module defines the structures and functions needed to encode network addresses in Bitcoin messages.

Structs

AddrV2Message

Address received from BIP155 addrv2 message

Address

A message which can be sent on the Bitcoin network

Enums

AddrV2

Supported networks for use in BIP155 addrv2 message