Module bitcoin::p2p

Bitcoin p2p network types.

This module defines support for (de)serialization and network transport of Bitcoin data and Bitcoin p2p network messages.

Re-exports

pub use self::address::Address;

std

Modules

addressstd

Bitcoin network addresses.

messagestd

Bitcoin network messages.

message_blockdatastd

Bitcoin blockdata network messages.

message_bloomstd

Bitcoin Connection Bloom filtering network messages.

message_compact_blocksstd

BIP152 Compact Blocks network messages

message_filterstd

Bitcoin Client Side Block Filtering network messages.

message_networkstd

Bitcoin network-related network messages.

Structs

Magic

Network magic bytes to identify the cryptocurrency network the message was intended for.

ParseMagicError

An error in parsing magic bytes.

ServiceFlags

Flags to indicate which network services a node supports.

UnknownMagicError

Error in creating a Network from Magic bytes.

Constants

PROTOCOL_VERSION

Version of the protocol as appearing in network message headers.