bitcoin

Module sign_message

Signature

This module provides signature related functions including secp256k1 signature recovery when library is used with the secp-recovery feature.

Structs

MessageSignature

A signature on a Bitcoin Signed Message.

Enums

MessageSignatureError

An error used for dealing with Bitcoin Signed Messages.

Constants

BITCOIN_SIGNED_MSG_PREFIX

The prefix for signed messages using Bitcoin’s message signing protocol.

Functions

signed_msg_hash

Hash message for signature using Bitcoin’s message signing format.