Module bitcoin::util

Utility functions

Functions needed by all parts of the Bitcoin library

Modules

address
Addresses

amount
Amounts

base58
Base58 encoder and decoder

bip32
BIP32 Implementation

bip143
BIP143 Implementation

bip158
BIP158 Compact Block Filters for Light Clients

contracthash	Deprecated
Pay-to-contract-hash support

ecdsa
ECDSA Bitcoin Keys

hash
Hash functions

key
Bitcoin Keys

merkleblock
Merkle Block and Partial Merkle Tree

misc
Miscellaneous functions

psbt
Partially Signed Transactions

schnorr
Schnorr Bitcoin Keys

taproot
Taproot

uint
Big unsigned integer types

Enums

Error
A general error code, other errors should implement conversions to/from this if appropriate.

Traits

BitArray
A trait which allows numbers to act as fixed-size bit arrays