Trait bitcoin::util::amount::serde::SerdeAmount

pub trait SerdeAmount: Copy + Sized {
    fn ser_sat<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>;

    fn des_sat<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>;

    fn ser_btc<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>;

    fn des_btc<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>;
}

This trait is used only to avoid code duplication and naming collisions of the different serde serialization crates.

TODO: Add the private::Sealed bound in next breaking release

Required methods

fn ser_sat<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>

fn des_sat<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>

fn ser_btc<S: Serializer>(self, s: S) -> Result<S::Ok, S::Error>

fn des_btc<'d, D: Deserializer<'d>>(d: D) -> Result<Self, D::Error>

Loading content...