Module bitcoin::util::bip152

BIP152 Compact Blocks

Implementation of compact blocks data structure and algorithms.

Structs

BlockTransactions

A BlockTransactions structure is used to provide some of the transactions in a block, as requested.

BlockTransactionsRequest

A BlockTransactionsRequest structure is used to list transaction indexes in a block being requested.

HeaderAndShortIds

A HeaderAndShortIds structure is used to relay a block header, the short transactions IDs used for matching already-available transactions, and a select few transactions which we expect a peer may be missing.

PrefilledTransaction

A PrefilledTransaction structure is used in HeaderAndShortIds to provide a list of a few transactions explicitly.

ShortId

Short transaction IDs are used to represent a transaction without sending a full 256-bit hash.

TxIndexOutOfRangeError

A transaction index is requested that is out of range from the corresponding block.

Enums

Error

A BIP-152 error