Struct bitcoin::util::bip32::DerivationPath

pub struct DerivationPath(_);

A BIP-32 derivation path.

Implementations

impl DerivationPath

pub fn len(&self) -> usize

Returns length of the derivation path

pub fn master() -> DerivationPath

Returns derivation path for a master key (i.e. empty derivation path)

pub fn is_master(&self) -> bool

Returns whether derivation path represents master key (i.e. it’s length is empty). True for m path.

pub fn child(&self, cn: ChildNumber) -> DerivationPath

Create a new DerivationPath that is a child of this one.

pub fn into_child(self, cn: ChildNumber) -> DerivationPath

Convert into a DerivationPath that is a child of this one.

pub fn children_from(&self, cn: ChildNumber) -> DerivationPathIterator<'_>

Get an Iterator over the children of this DerivationPath starting with the given ChildNumber.

pub fn normal_children(&self) -> DerivationPathIterator<'_>

Get an Iterator over the unhardened children of this DerivationPath.

pub fn hardened_children(&self) -> DerivationPathIterator<'_>

Get an Iterator over the hardened children of this DerivationPath.

pub fn extend<T: AsRef<[ChildNumber]>>(&self, path: T) -> DerivationPath

Concatenate self with path and return the resulting new path.

use bitcoin::util::bip32::{DerivationPath, ChildNumber};
use std::str::FromStr;

let base = DerivationPath::from_str("m/42").unwrap();

let deriv_1 = base.extend(DerivationPath::from_str("m/0/1").unwrap());
let deriv_2 = base.extend(&[
    ChildNumber::from_normal_idx(0).unwrap(),
    ChildNumber::from_normal_idx(1).unwrap()
]);

assert_eq!(deriv_1, deriv_2);