Module bitcoin::util::sighash

Generalized, efficient, signature hash implementation.

Implementation of the algorithm to compute the message to be signed according to Bip341, Bip143 and legacy (before Bip143).

Re-exports

pub use crate::blockdata::transaction::EcdsaSighashType;

pub use crate::blockdata::transaction::SighashTypeParseError;

Structs

Annex

The Annex struct is a slice wrapper enforcing first byte is 0x50.

ScriptPath

Information related to the script path spending.

SighashCache

Efficiently calculates signature hash message for legacy, segwit and taproot inputs.

Enums

Error

Possible errors in computing the signature message.

Prevouts

Contains outputs of previous transactions. In the case SchnorrSighashType variant is SIGHASH_ANYONECANPAY, Prevouts::One may be used.

SchnorrSighashType

Hashtype of an input’s signature, encoded in the last byte of the signature. Fixed values so they can be cast as integer types for encoding.