Trait bitcoin::util::BitArray

pub trait BitArray {
    fn bit(&self, idx: usize) -> bool;
    fn bit_slice(&self, start: usize, end: usize) -> Self;
    fn mask(&self, n: usize) -> Self;
    fn trailing_zeros(&self) -> usize;
    fn zero() -> Self;
    fn one() -> Self;
}

A trait which allows numbers to act as fixed-size bit arrays

Required Methods

fn bit(&self, idx: usize) -> bool

Is bit set?

fn bit_slice(&self, start: usize, end: usize) -> Self

Returns an array which is just the bits from start to end

fn mask(&self, n: usize) -> Self

Bitwise and with n ones

fn trailing_zeros(&self) -> usize

Trailing zeros

fn zero() -> Self

Create all-zeros value

fn one() -> Self

Create value representing one